lorawan-encoding: replace the generic-over-storage API in place
Deletes the T: AsRef<[u8]> parser and creator and moves the borrowed-view
API from the v2 staging module into their places: the views and owned
wire-field newtypes into parser.rs, the frame builders into creator.rs,
and the fallible MAC command framing (MacCommands, MacCommandSet,
ParseError) into maccommands.rs, with the certification and multicast
parse entry points beside their command enums. The MacCommandIterator
framing and the legacy parse_* functions are gone; lorawan-device (MAC
core, regions, and the whole async_device test suite) is migrated. Net
-3000 lines.

Validation leans on the previous commit, which kept both APIs side by
side and asserted byte-for-byte equivalence on the crate's shared test
vectors (parsing, building with MICs, join session-key derivation, MAC
command framing). With the old oracle gone, those assertions become
self-contained: creators still reproduce the wire vectors exactly,
decryption is pinned as an involution over the same vectors, and the
derived session keys are pinned to the captured values from that commit.

Migration notes beyond mechanical renames:

* The framing error is maccommands::ParseError, distinct from the
  payload-construction Error the module already had.
* McAddr becomes an owned wire-field newtype in parser.rs; the shared
  multicast payload types, key derivation and creators use it. Fhdr
  grows mc_addr() and DecryptedDataPayload::decrypt_in_place_with()
  takes raw AES keys for multicast session decryption.
* keys::DevEui/AppEui store [u8; 8] directly; parser::EUI64 is gone.
* The region trait's process_join_accept() takes Option<&CfList>
  instead of being generic over the decrypted join-accept buffer.
* TxFramesCtrlReqPayload::min_len() drops from 2 to 1. The legacy
  iterator compared the minimum against the data including the CID
  byte, so 2 effectively meant one payload byte; the new framing counts
  payload only, and the old value made it reject the spec's one-octet
  FrameType frame (caught by the device certification tests, which hung
  on it).
* The CommandHandler derive no longer emits the legacy Iterator impl.
* UnimplementedCreator moves to certification.rs, its only user.
* size-compare keeps the borrowed-view configurations only; the
  old-vs-new table stays in its README with a pointer at the previous
  commit.
