+++ lorawan-device/src/mac/mod.rs
@@ -165,12 +165,7 @@ impl Mac {
     }
 
     /// Join via ABP. This does not transmit a join request frame, but instead sets the session.
-    pub(crate) fn join_abp(
-        &mut self,
-        nwkskey: NwkSKey,
-        appskey: AppSKey,
-        devaddr: DevAddr<[u8; 4]>,
-    ) {
+    pub(crate) fn join_abp(&mut self, nwkskey: NwkSKey, appskey: AppSKey, devaddr: DevAddr) {
         self.state = State::Joined(Session::new(nwkskey, appskey, devaddr));
     }
