+++ lorawan-device/src/mac/multicast.rs
@@ -6,14 +6,14 @@ use core::fmt::Debug;
 use core::ops::RangeInclusive;
 use lorawan::default_crypto::DefaultFactory;
 use lorawan::keys::McKEKey;
+use lorawan::multicast::parse_downlink_multicast_commands;
 pub use lorawan::multicast::{self, Session};
 use lorawan::multicast::{
-    parse_downlink_multicast_messages, DownlinkRemoteSetup, McGroupDeleteAnsCreator,
-    McGroupSetupAnsCreator, McGroupStatusAnsCreator, PackageVersionAnsCreator,
+    DownlinkRemoteSetup, McGroupDeleteAnsCreator, McGroupSetupAnsCreator, McGroupStatusAnsCreator,
+    PackageVersionAnsCreator,
 };
-use lorawan::parser::FRMPayload;
 pub use lorawan::parser::McAddr;
-use lorawan::parser::{DataHeader, EncryptedDataPayload};
+use lorawan::parser::{DecryptedDataPayload, EncryptedDataPayload, FrmPayload};
 
 #[derive(Debug)]
 #[cfg_attr(feature = "defmt-03", derive(defmt::Format))]
@@ -65,8 +65,11 @@ impl Multicast {
     pub(crate) fn handle_rx<const D: usize>(
         &mut self,
         dl: &mut heapless::Vec<Downlink, D>,
-        encrypted_data: EncryptedDataPayload<&mut [u8]>,
+        bytes: &mut [u8],
     ) -> Response {
+        let Ok(encrypted_data) = EncryptedDataPayload::parse(bytes) else {
+            return Response::NoUpdate;
+        };
         let mc_addr = encrypted_data.fhdr().mc_addr();
         if let Some((group_id, session)) = self.matching_session(mc_addr) {
             let fcnt = encrypted_data.fhdr().fcnt() as u32;
@@ -76,19 +79,19 @@ impl Multicast {
                 return {
                     session.fcnt_down = fcnt;
                     // We can safely unwrap here because we already validated the MIC
-                    let decrypted = encrypted_data
-                        .decrypt(
-                            Some(session.mc_net_s_key().inner()),
-                            Some(session.mc_app_s_key().inner()),
-                            session.fcnt_down,
-                            &DefaultFactory,
-                        )
-                        .unwrap();
+                    let decrypted = DecryptedDataPayload::decrypt_in_place_with(
+                        bytes,
+                        Some(session.mc_net_s_key().inner()),
+                        Some(session.mc_app_s_key().inner()),
+                        session.fcnt_down,
+                        &DefaultFactory,
+                    )
+                    .unwrap();
                     if session.fcnt_down == session.max_fcnt_down() {
                         // if the FCnt is used up, the session has expired
                         Response::SessionExpired { group_id }
                     } else {
-                        if let (Some(fport), FRMPayload::Data(data)) =
+                        if let (Some(fport), FrmPayload::Data(data)) =
                             (decrypted.f_port(), decrypted.frm_payload())
                         {
                             // heapless Vec from slice fails only if slice is too large.
@@ -130,9 +133,12 @@ impl Multicast {
             return Response::NoUpdate;
         }
         let mc_k_e_key = self.mc_k_e_key.as_ref().unwrap();
-        let messages = parse_downlink_multicast_messages(data);
+        let messages = parse_downlink_multicast_commands(data);
         let mut new_session = None;
         for message in messages {
+            let Ok(message) = message else {
+                break;
+            };
             match message {
                 DownlinkRemoteSetup::McGroupSetupReq(mc_group_setup_req) => {
                     let crypto = DefaultFactory;
@@ -220,7 +226,7 @@ impl Multicast {
 
     pub(crate) fn matching_session(
         &mut self,
-        multicast_addr: McAddr<&[u8]>,
+        multicast_addr: McAddr,
     ) -> Option<(u8, &mut Session)> {
         self.sessions.iter_mut().enumerate().find_map(|(group_id, s)| {
             if let Some(s) = s {
