+++ lorawan-encoding/src/extra/mod.rs
@@ -1,3 +1,1 @@
 pub extern crate std;
-
-pub mod hasher;
