+++ lorawan-encoding/src/lib.rs
@@ -14,7 +14,6 @@ pub mod packet_length;
 pub mod parser;
 pub mod string;
 pub mod types;
-pub mod v2;
 
 #[cfg(feature = "full")]
 pub mod extra;
