+++ lorawan-encoding/src/maccommands.rs
@@ -167,55 +167,91 @@ macro_rules! create_value_reader_fn {
     )
 }
 
-/// Parses bytes to uplink MAC commands if possible.
+/// Error from parsing a MAC command stream.
 ///
-/// Could return error if some values are out of range or the payload does not end at MAC command
-/// boundry.
-/// # Argument
-///
-/// * bytes - the data from which the MAC commands are to be built.
-///
-/// # Examples
-///
-/// ```
-/// let mut data = vec![0x02, 0x03, 0x00];
-/// let mac_cmds: Vec<lorawan::maccommands::UplinkMacCommand> =
-///     lorawan::maccommands::parse_uplink_mac_commands(&data).collect();
-/// ```
-pub fn parse_uplink_mac_commands(data: &[u8]) -> MacCommandIterator<'_, UplinkMacCommand<'_>> {
-    MacCommandIterator::new(data)
-}
-/// Parses bytes to downlink MAC commands if possible.
-///
-/// Could return error if some values are out of range or the payload does not end at MAC command
-/// boundry.
-/// # Argument
-///
-/// * bytes - the data from which the MAC commands are to be built.
-///
-/// # Examples
-///
-/// ```
-/// let mut data = vec![0x02, 0x03, 0x00];
-/// let mac_cmds: Vec<lorawan::maccommands::DownlinkMacCommand> =
-///     lorawan::maccommands::parse_downlink_mac_commands(&data).collect();
-/// ```
-pub fn parse_downlink_mac_commands(data: &[u8]) -> MacCommandIterator<'_, DownlinkMacCommand<'_>> {
-    MacCommandIterator::new(data)
+/// Distinct from [`Error`], which covers individual payload construction:
+/// this one reports framing problems (the stream cannot be split into
+/// commands at all).
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+#[cfg_attr(feature = "defmt-03", derive(defmt::Format))]
+pub enum ParseError {
+    /// A CID that is not defined for this command set.
+    UnknownCid(u8),
+    /// The stream ends in the middle of this command's payload.
+    Truncated { cid: u8 },
 }
 
-/// Implementation of iterator for MAC commands.
-pub struct MacCommandIterator<'a, T> {
-    pub(crate) data: &'a [u8],
-    pub(crate) index: usize,
-    pub(crate) item: PhantomData<T>,
-}
-
-impl<'a, T> MacCommandIterator<'a, T> {
-    /// Creation.
-    pub fn new(data: &'a [u8]) -> Self {
-        Self { data, index: 0, item: Default::default() }
+impl core::fmt::Display for ParseError {
+    fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
+        match self {
+            ParseError::UnknownCid(cid) => write!(f, "unknown MAC command CID 0x{cid:02x}"),
+            ParseError::Truncated { cid } => {
+                write!(f, "truncated payload for MAC command CID 0x{cid:02x}")
+            }
+        }
     }
+}
+
+impl core::error::Error for ParseError {}
+
+/// One command set (a direction of a package): parses a single command off
+/// the front of a byte stream.
+///
+/// Implementations are generated by the `CommandHandler` derive macro from
+/// the enum's `#[cmd]` attributes.
+pub trait MacCommandSet<'a>: Sized {
+    /// Parses one command from the start of `data` (guaranteed non-empty by
+    /// the caller), returning the command and the number of bytes consumed.
+    fn parse_one(data: &'a [u8]) -> Result<(Self, usize), ParseError>;
+}
+
+/// Iterator over a MAC command stream, yielding `Result` per command.
+///
+/// Fused after the first error: a malformed tail yields exactly one `Err`.
+#[derive(Debug, Clone)]
+pub struct MacCommands<'a, T> {
+    data: &'a [u8],
+    errored: bool,
+    _commands: PhantomData<T>,
+}
+
+impl<'a, T> MacCommands<'a, T> {
+    #[inline]
+    pub fn new(data: &'a [u8]) -> Self {
+        Self { data, errored: false, _commands: PhantomData }
+    }
+}
+
+impl<'a, T: MacCommandSet<'a>> Iterator for MacCommands<'a, T> {
+    type Item = Result<T, ParseError>;
+
+    fn next(&mut self) -> Option<Self::Item> {
+        if self.errored || self.data.is_empty() {
+            return None;
+        }
+        match T::parse_one(self.data) {
+            Ok((cmd, consumed)) => {
+                self.data = &self.data[consumed..];
+                Some(Ok(cmd))
+            }
+            Err(e) => {
+                self.errored = true;
+                Some(Err(e))
+            }
+        }
+    }
+}
+
+/// Parses a stream of uplink (device-transmitted) MAC commands.
+#[inline]
+pub fn parse_uplink_mac_commands(data: &[u8]) -> MacCommands<'_, UplinkMacCommand<'_>> {
+    MacCommands::new(data)
+}
+
+/// Parses a stream of downlink (server-transmitted) MAC commands.
+#[inline]
+pub fn parse_downlink_mac_commands(data: &[u8]) -> MacCommands<'_, DownlinkMacCommand<'_>> {
+    MacCommands::new(data)
 }
 
 impl LinkCheckAnsPayload<'_> {
