+++ lorawan-encoding/src/parser.rs
@@ -1,1185 +1,1046 @@
-//! Provides types and methods for parsing LoRaWAN payloads.
+//! Zero-copy parsing of LoRaWAN frames as borrowed views.
+//!
+//! This module replaced the earlier `T: AsRef<[u8]>` generic-over-storage
+//! design:
+//!
+//! * Fixed-size wire fields ([`DevAddr`], [`DevEui`], [`JoinEui`],
+//!   [`DevNonce`], [`JoinNonce`], [`NetId`], [`Frequency`]) become plain
+//!   owned `[u8; N]` newtypes (`Copy`, `Eq`, `Hash`) with explicit
+//!   wire-order/value conversions.
+//! * Parsing produces zero-copy read-only views with all structural
+//!   validation done once at the boundary, so every accessor afterwards is
+//!   infallible by construction. One monomorphization regardless of where the
+//!   bytes live.
+//! * Decryption stays in place but takes the `&mut [u8]` explicitly, keeping
+//!   the encrypted/decrypted typestate without any storage generic.
+//! * MAC command parsing ([`crate::maccommands`], [`crate::certification`],
+//!   [`crate::multicast`]) yields `Result` per command instead of silently
+//!   stopping at malformed input,
+//!   with the framing generated by the `CommandHandler` derive from the
+//!   enums' `#[cmd]` attributes.
+//! * Frames are built from plain structs ([`DataFrame`], [`JoinRequest`],
+//!   [`JoinAccept`]) with `build_into(&mut buf)`; the types make most
+//!   invalid frames unrepresentable.
+//! * One [`Error`] enum implementing `Display` and `core::error::Error`.
+//!
+//! Deliberate divergences from the retired design, for anyone migrating:
+//!
+//! * `f_port()` returns `Some` when an FPort byte is present with an empty
+//!   FRMPayload. The old code reported `None` in that case, conflating
+//!   "FPort absent" with "FPort present, payload empty".
+//! * Wire fields are explicit about byte order: the wire transmits them
+//!   little-endian, so `from_value`/`value` convert accordingly. The old
+//!   `From<u32> for DevAddr` and `From<u16> for DevNonce` used big-endian
+//!   array order, which round-tripped through the wire with flipped
+//!   endianness.
+//! * All parse entry points reject frames with a non-R1 major version; the
+//!   old `EncryptedDataPayload::new` accepted them.
+//! * A MAC command stream with a malformed tail yields `Err`; the old
+//!   iterator ended silently.
 //!
 //! # Examples
 //!
-//! ```
-//! use lorawan::keys::*;
-//! use lorawan::parser::*;
+//! Receiving on a device: inspect the frame read-only, look up the session
+//! by [`DevAddr`], verify the MIC, then decrypt in place.
 //!
-//! let data = vec![
-//!     0x40, 0x04, 0x03, 0x02, 0x01, 0x80, 0x01, 0x00, 0x01, 0xa6, 0x94, 0x64, 0x26, 0x15, 0xd6,
-//!     0xc3, 0xb5, 0x82,
+//! ```
+//! use lorawan::default_crypto::DefaultFactory;
+//! use lorawan::keys::{AppSKey, NwkSKey};
+//! use lorawan::parser::{DecryptedDataPayload, DevAddr, EncryptedDataPayload, FrmPayload};
+//!
+//! let mut buf = [
+//!     0x40, 0x04, 0x03, 0x02, 0x01, 0x80, 0x01, 0x00, 0x01, 0xa6, 0x94, 0x64, 0x26, 0x15,
+//!     0xd6, 0xc3, 0xb5, 0x82,
 //! ];
-//! if let Ok(PhyPayload::Data(DataPayload::Encrypted(phy))) = parse(data) {
-//!     let key = AES128([1; 16]);
-//!     let decrypted = phy.decrypt(None, Some(&key), 1, &lorawan::default_crypto::DefaultFactory).unwrap();
-//!     if let FRMPayload::Data(data_payload) = decrypted.frm_payload() {
-//!         println!("{}", String::from_utf8_lossy(data_payload));
-//!     }
-//! } else {
-//!     panic!("failed to parse data payload");
-//! }
+//! let nwk_skey = NwkSKey::from([2; 16]);
+//! let app_skey = AppSKey::from([1; 16]);
+//!
+//! let phy = EncryptedDataPayload::parse(&buf).unwrap();
+//! assert_eq!(phy.fhdr().dev_addr(), DevAddr::from_value(0x01020304));
+//! assert!(phy.validate_mic(nwk_skey.inner(), 1, &DefaultFactory));
+//!
+//! let decrypted = DecryptedDataPayload::decrypt_in_place(
+//!     &mut buf,
+//!     Some(&nwk_skey),
+//!     Some(&app_skey),
+//!     1,
+//!     &DefaultFactory,
+//! )
+//! .unwrap();
+//! assert_eq!(decrypted.frm_payload(), FrmPayload::Data(b"hello"));
+//! ```
+//!
+//! Sending: describe the frame with a struct literal and build it into a
+//! caller-provided buffer.
+//!
+//! ```
+//! use core::num::NonZeroU8;
+//! use lorawan::default_crypto::DefaultFactory;
+//! use lorawan::keys::{AppSKey, NwkSKey};
+//! use lorawan::creator::{DataFrame, Payload};
+//! use lorawan::parser::{DataFrameType, DevAddr};
+//!
+//! let frame = DataFrame {
+//!     frame_type: DataFrameType::ConfirmedUp,
+//!     dev_addr: DevAddr::from_value(0x01020304),
+//!     adr: true,
+//!     fcnt: 42,
+//!     payload: Payload::Data { f_port: NonZeroU8::new(1).unwrap(), data: b"hello" },
+//!     ..Default::default()
+//! };
+//! let mut buf = [0u8; 64];
+//! let bytes = frame
+//!     .build_into(&mut buf, &NwkSKey::from([2; 16]), Some(&AppSKey::from([1; 16])), &DefaultFactory)
+//!     .unwrap();
+//! assert_eq!(bytes.len(), 18);
 //! ```
 
-use super::keys::{AppKey, AppSKey, CryptoFactory, Encrypter, NwkSKey, AES128, MIC};
-use crate::types::{ChannelMask, DLSettings, Frequency};
+use crate::keys::{AppKey, AppSKey, CryptoFactory, Encrypter, NwkSKey, AES128, MIC};
+use crate::packet_length::phy::join::{
+    JOIN_ACCEPT_LEN, JOIN_ACCEPT_WITH_CFLIST_LEN, JOIN_REQUEST_LEN,
+};
+use crate::packet_length::phy::{MHDR_LEN, MIC_LEN};
+use crate::securityhelpers;
+use crate::types::{ChannelMask, DLSettings};
 
-use super::securityhelpers;
-
-use super::packet_length::phy::{join::*, mac::FPORT_LEN, MHDR_LEN, MIC_LEN, PHY_PAYLOAD_MIN_LEN};
-
-#[derive(Debug, PartialEq)]
+/// Errors from parsing or decrypting a LoRaWAN frame.
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
 #[cfg_attr(feature = "defmt-03", derive(defmt::Format))]
 pub enum Error {
-    InvalidData,
-    InvalidMic,
-    InvalidKey,
-    InvalidMessageType,
-    InvalidPayload,
+    /// The buffer is shorter than the smallest possible frame.
+    TooShort,
+    /// The MHDR major version is not LoRaWAN R1.
     UnsupportedMajorVersion,
+    /// The MHDR message type is RFU or Proprietary.
+    UnsupportedMessageType,
+    /// The MHDR message type does not match the frame type being parsed.
+    UnexpectedMessageType,
+    /// The MHDR message type is not one of the four data frame types.
+    NotADataFrame,
+    /// The buffer length is not valid for this frame type.
+    InvalidLength,
+    /// FOptsLen claims more FOpts bytes than the frame contains.
+    TruncatedFhdr,
+    /// No key was supplied for the FPort carried by this frame.
+    MissingKey,
+    /// The MIC does not match the frame contents.
+    InvalidMic,
+    /// The output buffer is too small for the frame being built.
+    BufferTooShort,
+    /// FOpts is limited to 15 bytes.
+    FOptsTooLong,
+    /// The spec forbids FPort 0 (MAC commands in FRMPayload) together with a
+    /// non-empty FOpts.
+    FOptsWithFPortZero,
 }
 
-macro_rules! fixed_len_struct {
-    (
-        $(#[$outer:meta])*
-        struct $type:ident[$size:expr];
-    ) => {
-        $(#[$outer])*
-        #[derive(Debug, Eq)]
-        #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
-        #[cfg_attr(feature = "defmt-03", derive(defmt::Format))]
-        pub struct $type<T: AsRef<[u8]>>(T);
+impl core::fmt::Display for Error {
+    fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
+        let s = match self {
+            Error::BufferTooShort => "output buffer too small for the frame",
+            Error::FOptsTooLong => "FOpts is limited to 15 bytes",
+            Error::FOptsWithFPortZero => "FPort 0 with non-empty FOpts is forbidden",
+            Error::TooShort => "buffer shorter than minimal frame",
+            Error::UnsupportedMajorVersion => "unsupported LoRaWAN major version",
+            Error::UnsupportedMessageType => "RFU or proprietary MHDR message type",
+            Error::UnexpectedMessageType => "MHDR message type does not match frame type",
+            Error::NotADataFrame => "MHDR message type is not a data frame",
+            Error::InvalidLength => "buffer length invalid for this frame type",
+            Error::TruncatedFhdr => "FOptsLen exceeds frame length",
+            Error::MissingKey => "no key supplied for this frame's FPort",
+            Error::InvalidMic => "MIC mismatch",
+        };
+        f.write_str(s)
+    }
+}
 
+impl core::error::Error for Error {}
 
-        impl<T: AsRef<[u8]>> $type<T> {
-            pub const fn byte_len() -> usize {
-                $size
-            }
+/// A parsed LoRaWAN physical payload: a read-only view classified by MHDR.
+#[derive(Debug, Clone, Copy)]
+pub enum PhyPayload<'a> {
+    JoinRequest(JoinRequestPayload<'a>),
+    JoinAccept(EncryptedJoinAcceptPayload<'a>),
+    Data(EncryptedDataPayload<'a>),
+}
 
-            pub(crate) fn new_from_raw(bytes: T) -> $type<T> {
-                $type(bytes)
-            }
+/// Parses and classifies a LoRaWAN physical payload.
+#[inline]
+pub fn parse(bytes: &[u8]) -> Result<PhyPayload<'_>, Error> {
+    let [mhdr, ..] = bytes else {
+        return Err(Error::TooShort);
+    };
+    if mhdr & 0b11 != 0 {
+        return Err(Error::UnsupportedMajorVersion);
+    }
+    match mhdr >> 5 {
+        0 => Ok(PhyPayload::JoinRequest(JoinRequestPayload::parse(bytes)?)),
+        1 => Ok(PhyPayload::JoinAccept(EncryptedJoinAcceptPayload::parse(bytes)?)),
+        2..=5 => Ok(PhyPayload::Data(EncryptedDataPayload::parse(bytes)?)),
+        _ => Err(Error::UnsupportedMessageType),
+    }
+}
 
-            pub fn new(data: T) -> Option<$type<T>> {
-                let bytes = data.as_ref();
-                if bytes.len() != $size {
-                    None
-                } else {
-                    Some($type(data))
-                }
+// --- Data frames ---------------------------------------------------------
+
+/// The four data frame message types.
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+#[cfg_attr(feature = "defmt-03", derive(defmt::Format))]
+pub enum DataFrameType {
+    UnconfirmedUp,
+    UnconfirmedDown,
+    ConfirmedUp,
+    ConfirmedDown,
+}
+
+impl DataFrameType {
+    #[inline]
+    fn from_mhdr(mhdr: u8) -> Option<Self> {
+        match mhdr >> 5 {
+            2 => Some(Self::UnconfirmedUp),
+            3 => Some(Self::UnconfirmedDown),
+            4 => Some(Self::ConfirmedUp),
+            5 => Some(Self::ConfirmedDown),
+            _ => None,
+        }
+    }
+
+    #[inline]
+    pub fn is_uplink(self) -> bool {
+        matches!(self, Self::UnconfirmedUp | Self::ConfirmedUp)
+    }
+
+    #[inline]
+    pub fn is_confirmed(self) -> bool {
+        matches!(self, Self::ConfirmedUp | Self::ConfirmedDown)
+    }
+}
+
+/// Zero-copy view of the FHDR of a data frame.
+///
+/// Only constructed by [`EncryptedDataPayload::parse`] /
+/// [`DecryptedDataPayload`], which validate the length invariant
+/// (`bytes.len() == 7 + f_opts_len`), so accessors cannot panic.
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub struct Fhdr<'a> {
+    bytes: &'a [u8],
+    uplink: bool,
+}
+
+impl<'a> Fhdr<'a> {
+    #[inline]
+    pub fn dev_addr(&self) -> DevAddr {
+        DevAddr::from_wire_bytes(arr(&self.bytes[0..4]))
+    }
+
+    /// The address field interpreted as a multicast group address, for frames
+    /// received in a multicast session.
+    #[inline]
+    pub fn mc_addr(&self) -> McAddr {
+        McAddr::from_wire_bytes(arr(&self.bytes[0..4]))
+    }
+
+    #[inline]
+    pub fn fctrl(&self) -> FCtrl {
+        FCtrl::new(self.bytes[4], self.uplink)
+    }
+
+    /// The truncated 16-bit frame counter from the wire.
+    #[inline]
+    pub fn fcnt(&self) -> u16 {
+        u16::from_le_bytes([self.bytes[5], self.bytes[6]])
+    }
+
+    /// The FOpts bytes (may be empty). Length was validated at parse time.
+    #[inline]
+    pub fn f_opts(&self) -> &'a [u8] {
+        &self.bytes[7..]
+    }
+}
+
+/// Layout facts computed once during validation and shared by both payload
+/// views. Offsets are guaranteed in bounds for the buffer they were computed
+/// from.
+#[derive(Debug, Clone, Copy)]
+struct Layout {
+    frame_type: DataFrameType,
+    fhdr_len: usize,
+    /// Offset of the FPort byte, if present.
+    f_port_offset: Option<usize>,
+    /// FRMPayload byte range (empty if absent).
+    frm_start: usize,
+    frm_end: usize,
+}
+
+impl Layout {
+    /// Validates the structure of a data frame. This is the single place
+    /// bounds are checked; every view accessor relies on it.
+    #[inline]
+    fn validate(bytes: &[u8]) -> Result<Self, Error> {
+        // MHDR (1) + DevAddr (4) + FCtrl (1) + FCnt (2) + MIC (4)
+        const MIN_LEN: usize = MHDR_LEN + 7 + MIC_LEN;
+        if bytes.len() < MIN_LEN {
+            return Err(Error::TooShort);
+        }
+        let mhdr = bytes[0];
+        if mhdr & 0b11 != 0 {
+            return Err(Error::UnsupportedMajorVersion);
+        }
+        let frame_type = DataFrameType::from_mhdr(mhdr).ok_or(Error::NotADataFrame)?;
+        let fhdr_len = 7 + (bytes[5] & 0x0f) as usize;
+        let mic_offset = bytes.len() - MIC_LEN;
+        // FHDR must fit between MHDR and MIC.
+        if MHDR_LEN + fhdr_len > mic_offset {
+            return Err(Error::TruncatedFhdr);
+        }
+        let after_fhdr = MHDR_LEN + fhdr_len;
+        let (f_port_offset, frm_start) = if after_fhdr < mic_offset {
+            (Some(after_fhdr), after_fhdr + 1)
+        } else {
+            (None, after_fhdr)
+        };
+        Ok(Self { frame_type, fhdr_len, f_port_offset, frm_start, frm_end: mic_offset })
+    }
+}
+
+macro_rules! data_view_accessors {
+    () => {
+        #[inline]
+        pub fn frame_type(&self) -> DataFrameType {
+            self.layout.frame_type
+        }
+
+        #[inline]
+        pub fn is_uplink(&self) -> bool {
+            self.layout.frame_type.is_uplink()
+        }
+
+        #[inline]
+        pub fn is_confirmed(&self) -> bool {
+            self.layout.frame_type.is_confirmed()
+        }
+
+        #[inline]
+        pub fn fhdr(&self) -> Fhdr<'a> {
+            Fhdr {
+                bytes: &self.bytes[MHDR_LEN..MHDR_LEN + self.layout.fhdr_len],
+                uplink: self.is_uplink(),
             }
         }
 
-        impl<T: AsRef<[u8]> + Clone> Clone for $type<T> {
-            fn clone(&self) -> Self {
-                Self(self.0.clone())
-            }
+        /// The FPort, if the frame carries one.
+        ///
+        /// A present FPort with an empty FRMPayload is reported as `Some`
+        /// (the retired parser conflated it with no-FPort).
+        #[inline]
+        pub fn f_port(&self) -> Option<u8> {
+            self.layout.f_port_offset.map(|off| self.bytes[off])
         }
 
-        impl<T: AsRef<[u8]> + Copy> Copy for $type<T> {
+        #[inline]
+        pub fn mic(&self) -> MIC {
+            let mic = &self.bytes[self.bytes.len() - MIC_LEN..];
+            MIC([mic[0], mic[1], mic[2], mic[3]])
         }
 
-        impl<T: AsRef<[u8]>, V: AsRef<[u8]>> PartialEq<$type<T>> for $type<V> {
-            fn eq(&self, other: &$type<T>) -> bool {
-                self.as_ref() == other.as_ref()
-            }
+        /// The raw frame bytes.
+        #[inline]
+        pub fn as_bytes(&self) -> &'a [u8] {
+            self.bytes
         }
-
-        impl<'a> From<&'a [u8; $size]> for $type<&'a [u8; $size]> {
-            fn from(v: &'a [u8; $size]) -> Self {
-                $type(v)
-            }
-        }
-
-        impl From<[u8; $size]> for $type<[u8; $size]> {
-            fn from(v: [u8; $size]) -> Self {
-                $type(v)
-            }
-        }
-
-        impl<T: AsRef<[u8]>> AsRef<[u8]> for $type<T> {
-            fn as_ref(&self) -> &[u8] {
-                self.0.as_ref()
-            }
-        }
-
-        impl<T: AsRef<[u8]>> $type<T> {
-            #[inline]
-            pub fn to_owned(&self) -> $type<[u8; $size]> {
-                let mut data = [0 as u8; $size];
-                data.copy_from_slice(self.0.as_ref());
-                $type(data)
-            }
-        }
-
-        impl<T: AsRef<[u8]> + Default> Default for $type<T> {
-            #[inline]
-            fn default() -> $type<T> {
-                $type(T::default())
-            }
-        }
-
     };
 }
 
-/// PhyPayload is a type that represents a physical LoRaWAN payload.
+/// Zero-copy read-only view of an encrypted data frame.
 ///
-/// It can either be JoinRequest, JoinAccept, or DataPayload.
-#[derive(Debug, PartialEq, Eq)]
-pub enum PhyPayload<T> {
-    JoinRequest(JoinRequestPayload<T>),
-    JoinAccept(JoinAcceptPayload<T>),
-    Data(DataPayload<T>),
+/// One monomorphization regardless of where the bytes live; inspect the
+/// header, look up session keys by [`DevAddr`], check the MIC, then hand the
+/// mutable buffer to [`DecryptedDataPayload::decrypt_in_place`].
+#[derive(Debug, Clone, Copy)]
+pub struct EncryptedDataPayload<'a> {
+    bytes: &'a [u8],
+    layout: Layout,
 }
 
-#[cfg(feature = "defmt-03")]
-impl<T: defmt::Format> defmt::Format for PhyPayload<T> {
-    fn format(&self, f: defmt::Formatter<'_>) {
-        match self {
-            PhyPayload::JoinRequest(r) => {
-                defmt::write!(f, "JoinRequestPayload({})", r.0);
-            }
-            PhyPayload::JoinAccept(a) => match a {
-                JoinAcceptPayload::Encrypted(data) => {
-                    defmt::write!(f, "JoinAcceptPayload::Encrypted({})", data.0);
-                }
-                JoinAcceptPayload::Decrypted(data) => {
-                    defmt::write!(f, "JoinAcceptPayload::Decrypted({})", data.0);
-                }
-            },
-            PhyPayload::Data(d) => match d {
-                DataPayload::Encrypted(data) => {
-                    defmt::write!(f, "DataPayload::Encrypted({})", data.0);
-                }
-                DataPayload::Decrypted(data) => {
-                    defmt::write!(f, "DataPayload::Decrypted({})", data.0);
-                }
-            },
-        };
+impl<'a> EncryptedDataPayload<'a> {
+    /// Parses and structurally validates an encrypted data frame.
+    #[inline]
+    pub fn parse(bytes: &'a [u8]) -> Result<Self, Error> {
+        let layout = Layout::validate(bytes)?;
+        Ok(Self { bytes, layout })
+    }
+
+    data_view_accessors!();
+
+    /// Whether the MIC matches under `key` and the given 32-bit frame counter.
+    #[inline]
+    pub fn validate_mic<C: CryptoFactory>(&self, key: &AES128, fcnt: u32, crypto: &C) -> bool {
+        let without_mic = &self.bytes[..self.bytes.len() - MIC_LEN];
+        self.mic() == securityhelpers::calculate_data_mic(without_mic, crypto.new_mac(key), fcnt)
     }
 }
 
-impl<T: AsRef<[u8]>> AsRef<[u8]> for PhyPayload<T> {
-    fn as_ref(&self) -> &[u8] {
-        match self {
-            PhyPayload::JoinRequest(jr) => jr.as_bytes(),
-            PhyPayload::JoinAccept(ja) => ja.as_bytes(),
-            PhyPayload::Data(data) => data.as_bytes(),
-        }
-    }
-}
-
-/// JoinAcceptPayload is a type that represents a JoinAccept.
+/// A decrypted data frame, borrowing the buffer that was decrypted in place.
 ///
-/// It can either be encrypted for example as a result from the [parse](fn.parse.html)
-/// function or not.
+/// Only obtainable through [`Self::decrypt_in_place`] or
+/// [`Self::check_mic_and_decrypt_in_place`], which preserves the
+/// encrypted/decrypted typestate of the old design without a storage generic.
+#[derive(Debug)]
+pub struct DecryptedDataPayload<'a> {
+    bytes: &'a [u8],
+    layout: Layout,
+}
+
+/// The decrypted FRMPayload contents.
 #[derive(Debug, PartialEq, Eq)]
-pub enum JoinAcceptPayload<T> {
-    Encrypted(EncryptedJoinAcceptPayload<T>),
-    Decrypted(DecryptedJoinAcceptPayload<T>),
+pub enum FrmPayload<'a> {
+    /// Application data (FPort > 0).
+    Data(&'a [u8]),
+    /// MAC commands (FPort == 0).
+    MacCommands(&'a [u8]),
+    /// The frame carries no FRMPayload.
+    None,
 }
 
-impl<T: AsRef<[u8]>> AsPhyPayloadBytes for JoinAcceptPayload<T> {
-    fn as_bytes(&self) -> &[u8] {
-        match self {
-            JoinAcceptPayload::Encrypted(e) => e.as_bytes(),
-            JoinAcceptPayload::Decrypted(d) => d.as_bytes(),
-        }
-    }
-}
-
-/// DataPayload is a type that represents a ConfirmedDataUp, ConfirmedDataDown,
-/// UnconfirmedDataUp or UnconfirmedDataDown.
-///
-/// It can either be encrypted for example as a result from the [parse](fn.parse.html)
-/// function or not.
-#[derive(Debug, PartialEq, Eq)]
-pub enum DataPayload<T> {
-    Encrypted(EncryptedDataPayload<T>),
-    Decrypted(DecryptedDataPayload<T>),
-}
-
-impl<T: AsRef<[u8]>> DataHeader for DataPayload<T> {
-    fn as_data_bytes(&self) -> &[u8] {
-        match self {
-            DataPayload::Encrypted(data) => data.as_data_bytes(),
-            DataPayload::Decrypted(data) => data.as_data_bytes(),
-        }
-    }
-}
-
-/// Trait with the sole purpose to make clear distinction in some implementations between types
-/// that just happen to have AsRef and those that want to have the given implementations (like
-/// MICAble and MHDRAble).
-pub trait AsPhyPayloadBytes {
-    fn as_bytes(&self) -> &[u8];
-}
-
-impl AsRef<[u8]> for dyn AsPhyPayloadBytes {
-    fn as_ref(&self) -> &[u8] {
-        self.as_bytes()
-    }
-}
-
-/// Helper trait to add mic to all types that should have it.
-pub trait MICAble {
-    /// Gives the MIC of the PhyPayload.
-    fn mic(&self) -> MIC;
-}
-
-impl<T: AsPhyPayloadBytes> MICAble for T {
-    fn mic(&self) -> MIC {
-        let data = self.as_bytes();
-        let len = data.len();
-        MIC([data[len - 4], data[len - 3], data[len - 2], data[len - 1]])
-    }
-}
-
-/// Helper trait to add mhdr to all types that should have it.
-pub trait MHDRAble {
-    /// Gives the MIC of the PhyPayload.
-    fn mhdr(&self) -> MHDR;
-}
-
-/// Assumes at least one byte in the data.
-impl<T: AsPhyPayloadBytes> MHDRAble for T {
-    fn mhdr(&self) -> MHDR {
-        let data = self.as_bytes();
-        MHDR(data[0])
-    }
-}
-
-/// JoinAcceptPayload represents a JoinRequest.
-///
-/// It can be built either directly through the [new](#method.new) or using the
-/// [parse](fn.parse.html) function.
-#[derive(Debug, PartialEq, Eq)]
-pub struct JoinRequestPayload<T>(T);
-
-impl<T: AsRef<[u8]>> AsPhyPayloadBytes for JoinRequestPayload<T> {
-    fn as_bytes(&self) -> &[u8] {
-        self.0.as_ref()
-    }
-}
-
-impl<T: AsRef<[u8]>> JoinRequestPayload<T> {
-    /// Creates a new JoinRequestPayload if the provided data is acceptable.
+impl<'a> DecryptedDataPayload<'a> {
+    /// Parses `buf` and decrypts its FRMPayload in place.
     ///
-    /// # Argument
+    /// Does NOT verify the MIC; prefer [`Self::check_mic_and_decrypt_in_place`]
+    /// unless the MIC was already checked via
+    /// [`EncryptedDataPayload::validate_mic`].
     ///
-    /// * data - the bytes for the payload.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// let data = vec![
-    ///     0x00, 0x04, 0x03, 0x02, 0x01, 0x04, 0x03, 0x02, 0x01, 0x05, 0x04, 0x03, 0x02, 0x05, 0x04,
-    ///     0x03, 0x02, 0x2d, 0x10, 0x6a, 0x99, 0x0e, 0x12,
-    /// ];
-    /// let phy = lorawan::parser::JoinRequestPayload::new(
-    ///     data
-    /// );
-    /// ```
-    pub fn new(data: T) -> Result<Self, Error> {
-        if !Self::can_build_from(data.as_ref()) {
-            Err(Error::InvalidData)
-        } else {
-            Ok(Self(data))
-        }
-    }
-
-    fn can_build_from(bytes: &[u8]) -> bool {
-        bytes.len() == JOIN_REQUEST_LEN && MHDR(bytes[0]).mtype() == MType::JoinRequest
-    }
-
-    /// Gives the APP EUI of the JoinRequest.
-    pub fn app_eui(&self) -> EUI64<&[u8]> {
-        EUI64::new_from_raw(&self.0.as_ref()[1..9])
-    }
-
-    /// Gives the DEV EUI of the JoinRequest.
-    pub fn dev_eui(&self) -> EUI64<&[u8]> {
-        EUI64::new_from_raw(&self.0.as_ref()[9..17])
-    }
-
-    /// Gives the DEV Nonce of the JoinRequest.
-    pub fn dev_nonce(&self) -> DevNonce<&[u8]> {
-        DevNonce::new_from_raw(&self.0.as_ref()[17..19])
-    }
-
-    /// Verifies that the JoinRequest has correct MIC.
-    pub fn validate_mic<C: CryptoFactory>(&self, key: &AES128, crypto: &C) -> bool {
-        self.mic() == self.calculate_mic(key, crypto)
-    }
-
-    fn calculate_mic<C: CryptoFactory>(&self, key: &AES128, crypto: &C) -> MIC {
-        let d = self.0.as_ref();
-        securityhelpers::calculate_mic(&d[..d.len() - MIC_LEN], crypto.new_mac(key))
-    }
-}
-
-/// EncryptedJoinAcceptPayload represents an encrypted JoinAccept.
-///
-/// It can be built either directly through the [new](#method.new) or using the
-/// [parse](fn.parse.html) function.
-#[derive(Debug, PartialEq, Eq)]
-pub struct EncryptedJoinAcceptPayload<T>(T);
-
-impl<T: AsRef<[u8]>> AsPhyPayloadBytes for EncryptedJoinAcceptPayload<T> {
-    fn as_bytes(&self) -> &[u8] {
-        self.0.as_ref()
-    }
-}
-
-impl<T: AsRef<[u8]> + AsMut<[u8]>> EncryptedJoinAcceptPayload<T> {
-    /// Creates a new EncryptedJoinAcceptPayload if the provided data is acceptable.
-    ///
-    /// # Argument
-    ///
-    /// * data - the bytes for the payload.
-    /// * factory - the factory that shall be used to create object for crypto functions.
-    pub fn new(data: T) -> Result<Self, Error> {
-        if Self::can_build_from(data.as_ref()) {
-            Ok(Self(data))
-        } else {
-            Err(Error::InvalidData)
-        }
-    }
-
-    fn can_build_from(bytes: &[u8]) -> bool {
-        (bytes.len() == JOIN_ACCEPT_LEN || bytes.len() == JOIN_ACCEPT_WITH_CFLIST_LEN)
-            && MHDR(bytes[0]).mtype() == MType::JoinAccept
-    }
-
-    /// Decrypts the EncryptedJoinAcceptPayload producing a DecryptedJoinAcceptPayload.
-    ///
-    /// This method consumes the EncryptedJoinAcceptPayload as it reuses the underlying memory.
-    /// Please note that it does not verify the mic.
-    ///
-    /// # Argument
-    ///
-    /// * key - the key to be used for the decryption.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// let mut data = vec![
-    ///     0x20, 0x49, 0x3e, 0xeb, 0x51, 0xfb, 0xa2, 0x11, 0x6f, 0x81, 0x0e, 0xdb, 0x37, 0x42, 0x97,
-    ///     0x51, 0x42,
-    /// ];
-    /// let phy = lorawan::parser::EncryptedJoinAcceptPayload::new(data);
-    /// let key = lorawan::keys::AppKey::from([
-    ///     0x00, 0x11, 0x22, 0x33, 0x44, 0x55, 0x66, 0x77, 0x88, 0x99, 0xaa, 0xbb, 0xcc, 0xdd, 0xee,
-    ///     0xff,
-    /// ]);
-    /// let decrypted = phy.unwrap().decrypt(&key,&lorawan::default_crypto::DefaultFactory);
-    /// ```
-    pub fn decrypt<C: CryptoFactory>(
-        mut self,
-        key: &AppKey,
+    /// Key selection follows the spec: `app_skey` decrypts FPort > 0,
+    /// `nwk_skey` decrypts FPort 0 (MAC commands). Only the key the frame
+    /// actually needs must be present. `fcnt` supplies the upper 16 bits of
+    /// the frame counter; the lower 16 come from the wire.
+    #[inline]
+    pub fn decrypt_in_place<C: CryptoFactory>(
+        buf: &'a mut [u8],
+        nwk_skey: Option<&NwkSKey>,
+        app_skey: Option<&AppSKey>,
+        fcnt: u32,
         crypto: &C,
-    ) -> DecryptedJoinAcceptPayload<T> {
-        {
-            let bytes = self.0.as_mut();
-            let len = bytes.len();
-            let aes_enc = crypto.new_enc(&key.0);
+    ) -> Result<Self, Error> {
+        Self::decrypt_in_place_with(
+            buf,
+            nwk_skey.map(|k| k.inner()),
+            app_skey.map(|k| k.inner()),
+            fcnt,
+            crypto,
+        )
+    }
 
-            for i in 0..(len >> 4) {
-                let start = (i << 4) + 1;
-                aes_enc.encrypt_block(&mut bytes[start..(start + 16)]);
-            }
+    /// Like [`Self::decrypt_in_place`], but takes raw AES keys. For sessions
+    /// whose keys are not the unicast [`NwkSKey`]/[`AppSKey`] pair, such as
+    /// multicast sessions (`McNetSKey`/`McAppSKey`).
+    #[inline]
+    pub fn decrypt_in_place_with<C: CryptoFactory>(
+        buf: &'a mut [u8],
+        nwk_key: Option<&AES128>,
+        app_key: Option<&AES128>,
+        fcnt: u32,
+        crypto: &C,
+    ) -> Result<Self, Error> {
+        let layout = Layout::validate(buf)?;
+        if layout.frm_start < layout.frm_end {
+            let uses_app_key = matches!(layout.f_port_offset, Some(off) if buf[off] != 0);
+            let key = if uses_app_key {
+                app_key
+            } else {
+                nwk_key
+            };
+            let key = key.ok_or(Error::MissingKey)?;
+            let wire_fcnt = u16::from_le_bytes([buf[6], buf[7]]);
+            let full_fcnt = ((fcnt >> 16) << 16) | u32::from(wire_fcnt);
+            securityhelpers::encrypt_frm_data_payload(
+                buf,
+                layout.frm_start,
+                layout.frm_end,
+                full_fcnt,
+                &crypto.new_enc(key),
+            );
         }
-        DecryptedJoinAcceptPayload(self.0)
+        Ok(Self { bytes: buf, layout })
+    }
+
+    /// Verifies the MIC with `nwk_skey`, then decrypts in place.
+    ///
+    /// The MIC is checked before anything is written, so on `InvalidMic` the
+    /// buffer still holds the received bytes unchanged.
+    #[inline]
+    pub fn check_mic_and_decrypt_in_place<C: CryptoFactory>(
+        buf: &'a mut [u8],
+        nwk_skey: &NwkSKey,
+        app_skey: Option<&AppSKey>,
+        fcnt: u32,
+        crypto: &C,
+    ) -> Result<Self, Error> {
+        if !EncryptedDataPayload::parse(buf)?.validate_mic(nwk_skey.inner(), fcnt, crypto) {
+            return Err(Error::InvalidMic);
+        }
+        Self::decrypt_in_place(buf, Some(nwk_skey), app_skey, fcnt, crypto)
+    }
+
+    data_view_accessors!();
+
+    /// The decrypted FRMPayload.
+    #[inline]
+    pub fn frm_payload(&self) -> FrmPayload<'a> {
+        let frm = &self.bytes[self.layout.frm_start..self.layout.frm_end];
+        match self.f_port() {
+            None => FrmPayload::None,
+            Some(0) => FrmPayload::MacCommands(frm),
+            Some(_) => FrmPayload::Data(frm),
+        }
     }
 }
 
-/// DecryptedJoinAcceptPayload represents a decrypted JoinAccept.
+// --- JoinRequest / JoinAccept --------------------------------------------
+
+#[inline]
+fn check_mhdr(bytes: &[u8], mtype: u8) -> Result<(), Error> {
+    let [mhdr, ..] = bytes else {
+        return Err(Error::TooShort);
+    };
+    if mhdr & 0b11 != 0 {
+        return Err(Error::UnsupportedMajorVersion);
+    }
+    if mhdr >> 5 != mtype {
+        return Err(Error::UnexpectedMessageType);
+    }
+    Ok(())
+}
+
+#[inline]
+fn extract_mic(bytes: &[u8]) -> MIC {
+    MIC(arr(&bytes[bytes.len() - MIC_LEN..]))
+}
+
+/// Zero-copy view of a JoinRequest frame.
 ///
-/// It can be built either directly through the [new](#method.new) or using the
-/// [EncryptedJoinAcceptPayload.decrypt](struct.EncryptedJoinAcceptPayload.html#method.decrypt)
-/// function.
-#[derive(Debug, PartialEq, Eq)]
-pub struct DecryptedJoinAcceptPayload<T>(T);
-
-impl<T: AsRef<[u8]>> AsPhyPayloadBytes for DecryptedJoinAcceptPayload<T> {
-    fn as_bytes(&self) -> &[u8] {
-        self.0.as_ref()
-    }
+/// The fixed 23-byte length is checked once at parse time; the fixed-size
+/// reference makes every accessor's bounds visible to the compiler.
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub struct JoinRequestPayload<'a> {
+    bytes: &'a [u8; JOIN_REQUEST_LEN],
 }
 
-impl<T: AsRef<[u8]>> DecryptedJoinAcceptPayload<T> {
-    /// Verifies that the JoinAccept has correct MIC.
+impl<'a> JoinRequestPayload<'a> {
+    /// Parses and structurally validates a JoinRequest.
+    #[inline]
+    pub fn parse(bytes: &'a [u8]) -> Result<Self, Error> {
+        check_mhdr(bytes, 0)?;
+        let bytes = bytes.try_into().map_err(|_| Error::InvalidLength)?;
+        Ok(Self { bytes })
+    }
+
+    /// The JoinEUI (called AppEUI before LoRaWAN 1.0.4).
+    #[inline]
+    pub fn join_eui(&self) -> JoinEui {
+        JoinEui::from_wire_bytes(arr(&self.bytes[1..9]))
+    }
+
+    #[inline]
+    pub fn dev_eui(&self) -> DevEui {
+        DevEui::from_wire_bytes(arr(&self.bytes[9..17]))
+    }
+
+    #[inline]
+    pub fn dev_nonce(&self) -> DevNonce {
+        DevNonce::from_wire_bytes(arr(&self.bytes[17..19]))
+    }
+
+    #[inline]
+    pub fn mic(&self) -> MIC {
+        extract_mic(self.bytes)
+    }
+
+    /// Whether the MIC matches under the device's root key.
+    #[inline]
     pub fn validate_mic<C: CryptoFactory>(&self, key: &AppKey, crypto: &C) -> bool {
-        self.mic() == self.calculate_mic(key, crypto)
+        let without_mic = &self.bytes[..JOIN_REQUEST_LEN - MIC_LEN];
+        self.mic() == securityhelpers::calculate_mic(without_mic, crypto.new_mac(key.inner()))
     }
 
-    pub fn calculate_mic<C: CryptoFactory>(&self, key: &AppKey, crypto: &C) -> MIC {
-        let d = self.0.as_ref();
-        securityhelpers::calculate_mic(&d[..d.len() - MIC_LEN], crypto.new_mac(&key.0))
+    /// The raw frame bytes.
+    #[inline]
+    pub fn as_bytes(&self) -> &'a [u8] {
+        self.bytes
+    }
+}
+
+#[inline]
+fn validate_join_accept_structure(bytes: &[u8]) -> Result<(), Error> {
+    check_mhdr(bytes, 1)?;
+    if bytes.len() != JOIN_ACCEPT_LEN && bytes.len() != JOIN_ACCEPT_WITH_CFLIST_LEN {
+        return Err(Error::InvalidLength);
+    }
+    Ok(())
+}
+
+/// Zero-copy view of a still-encrypted JoinAccept frame.
+///
+/// Nothing but the MHDR is readable before decryption; the payload (including
+/// its MIC) is encrypted with the AppKey. Hand the mutable buffer to
+/// [`DecryptedJoinAcceptPayload::decrypt_in_place`] to proceed.
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub struct EncryptedJoinAcceptPayload<'a> {
+    bytes: &'a [u8],
+}
+
+impl<'a> EncryptedJoinAcceptPayload<'a> {
+    /// Parses and structurally validates an encrypted JoinAccept.
+    #[inline]
+    pub fn parse(bytes: &'a [u8]) -> Result<Self, Error> {
+        validate_join_accept_structure(bytes)?;
+        Ok(Self { bytes })
     }
 
-    /// Computes the network session key for a given device.
+    /// The raw frame bytes.
+    #[inline]
+    pub fn as_bytes(&self) -> &'a [u8] {
+        self.bytes
+    }
+}
+
+/// A decrypted JoinAccept, borrowing the buffer that was decrypted in place.
+#[derive(Debug)]
+pub struct DecryptedJoinAcceptPayload<'a> {
+    bytes: &'a [u8],
+}
+
+/// The optional channel frequency list of a JoinAccept, fully owned.
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub enum CfList {
+    /// CFList type 0: five additional channel frequencies.
+    DynamicChannel([Frequency; 5]),
+    /// CFList type 1: a 72-entry channel mask for fixed-plan regions.
+    FixedChannel(ChannelMask<9>),
+}
+
+impl<'a> DecryptedJoinAcceptPayload<'a> {
+    /// Parses `buf` and decrypts it in place.
     ///
-    /// # Argument
+    /// Does NOT verify the MIC (it only becomes readable after decryption);
+    /// prefer [`Self::check_mic_and_decrypt_in_place`].
+    #[inline]
+    pub fn decrypt_in_place<C: CryptoFactory>(
+        buf: &'a mut [u8],
+        key: &AppKey,
+        crypto: &C,
+    ) -> Result<Self, Error> {
+        validate_join_accept_structure(buf)?;
+        // The device side runs AES *encryption* to invert the server's
+        // decrypt-mode transformation, per the spec.
+        let aes = crypto.new_enc(key.inner());
+        for block in buf[1..].chunks_exact_mut(16) {
+            aes.encrypt_block(block);
+        }
+        Ok(Self { bytes: buf })
+    }
+
+    /// Decrypts in place, then verifies the MIC.
     ///
-    /// * app_nonce - the network server nonce.
-    /// * nwk_addr - the address of the network.
-    /// * dev_nonce - the nonce from the device.
-    /// * key - the app key.
+    /// Because the MIC only becomes readable after decryption, `buf` has
+    /// already been transformed when `InvalidMic` is returned; with the wrong
+    /// key that leaves it garbage, not the received bytes. Unlike the data
+    /// frame path, a failed authentication here cannot leave the buffer
+    /// untouched. Callers that want to retry with another key or log the
+    /// original frame must keep a copy.
+    #[inline]
+    pub fn check_mic_and_decrypt_in_place<C: CryptoFactory>(
+        buf: &'a mut [u8],
+        key: &AppKey,
+        crypto: &C,
+    ) -> Result<Self, Error> {
+        let decrypted = Self::decrypt_in_place(buf, key, crypto)?;
+        if !decrypted.validate_mic(key, crypto) {
+            return Err(Error::InvalidMic);
+        }
+        Ok(decrypted)
+    }
+
+    /// Whether the MIC matches under the device's root key.
+    #[inline]
+    pub fn validate_mic<C: CryptoFactory>(&self, key: &AppKey, crypto: &C) -> bool {
+        let without_mic = &self.bytes[..self.bytes.len() - MIC_LEN];
+        self.mic() == securityhelpers::calculate_mic(without_mic, crypto.new_mac(key.inner()))
+    }
+
+    /// The server nonce (called AppNonce before LoRaWAN 1.0.4).
+    #[inline]
+    pub fn join_nonce(&self) -> JoinNonce {
+        JoinNonce::from_wire_bytes(arr(&self.bytes[1..4]))
+    }
+
+    #[inline]
+    pub fn net_id(&self) -> NetId {
+        NetId::from_wire_bytes(arr(&self.bytes[4..7]))
+    }
+
+    #[inline]
+    pub fn dev_addr(&self) -> DevAddr {
+        DevAddr::from_wire_bytes(arr(&self.bytes[7..11]))
+    }
+
+    #[inline]
+    pub fn dl_settings(&self) -> DLSettings {
+        DLSettings::new(self.bytes[11])
+    }
+
+    /// The RX1 delay in seconds (1..=15; the wire value 0 means 1).
+    #[inline]
+    pub fn rx_delay(&self) -> u8 {
+        self.bytes[12] & 0x0f
+    }
+
+    /// The channel frequency list, when present and of a known type.
     ///
-    /// # Examples
-    ///
-    /// ```
-    /// let dev_nonce = vec![0xcc, 0xdd];
-    /// let data = vec![
-    ///     0x20, 0x49, 0x3e, 0xeb, 0x51, 0xfb, 0xa2, 0x11, 0x6f, 0x81, 0x0e, 0xdb, 0x37, 0x42, 0x97,
-    ///     0x51, 0x42,
-    /// ];
-    /// let app_key = lorawan::keys::AppKey::from([
-    ///     0x00, 0x11, 0x22, 0x33, 0x44, 0x55, 0x66, 0x77, 0x88, 0x99, 0xaa, 0xbb, 0xcc, 0xdd, 0xee,
-    ///     0xff,
-    /// ]);
-    /// let join_accept = lorawan::parser::EncryptedJoinAcceptPayload::new(data).unwrap().decrypt(&app_key, &lorawan::default_crypto::DefaultFactory);
-    ///
-    /// let nwk_skey = join_accept
-    ///     .derive_nwkskey(&lorawan::parser::DevNonce::new(&dev_nonce[..]).unwrap(), &app_key, &lorawan::default_crypto::DefaultFactory);
-    /// ```
-    pub fn derive_nwkskey<TT: AsRef<[u8]>, C: CryptoFactory>(
+    /// Returns `None` both when the frame carries no CFList and when the
+    /// CFList type octet is RFU, as the retired parser did.
+    #[inline]
+    pub fn c_f_list(&self) -> Option<CfList> {
+        if self.bytes.len() == JOIN_ACCEPT_LEN {
+            return None;
+        }
+        let cflist = &self.bytes[13..29];
+        match cflist[15] {
+            0 => {
+                let mut freqs = [Frequency::default(); 5];
+                for (freq, chunk) in freqs.iter_mut().zip(cflist.chunks_exact(3)) {
+                    *freq = Frequency::from_wire_bytes(arr(chunk));
+                }
+                Some(CfList::DynamicChannel(freqs))
+            }
+            1 => Some(CfList::FixedChannel(ChannelMask::new_from_raw(&cflist[..9]))),
+            _ => None,
+        }
+    }
+
+    #[inline]
+    pub fn mic(&self) -> MIC {
+        extract_mic(self.bytes)
+    }
+
+    /// The raw (decrypted) frame bytes.
+    #[inline]
+    pub fn as_bytes(&self) -> &'a [u8] {
+        self.bytes
+    }
+
+    /// Derives the network session key for this join.
+    #[inline]
+    pub fn derive_nwkskey<C: CryptoFactory>(
         &self,
-        dev_nonce: &DevNonce<TT>,
+        dev_nonce: DevNonce,
         key: &AppKey,
         crypto: &C,
     ) -> NwkSKey {
-        NwkSKey(self.derive_session_key(0x1, dev_nonce, &key.0, crypto))
+        NwkSKey(self.derive_session_key(0x01, dev_nonce, key, crypto))
     }
 
-    #[deprecated(since = "0.9.1", note = "Please use `derive_nwkskey` instead")]
-    pub fn derive_newskey<TT: AsRef<[u8]>, C: CryptoFactory>(
+    /// Derives the application session key for this join.
+    #[inline]
+    pub fn derive_appskey<C: CryptoFactory>(
         &self,
-        dev_nonce: &DevNonce<TT>,
-        key: &AppKey,
-        crypto: &C,
-    ) -> NwkSKey {
-        NwkSKey(self.derive_session_key(0x1, dev_nonce, &key.0, crypto))
-    }
-
-    /// Computes the application session key for a given device.
-    ///
-    /// # Argument
-    ///
-    /// * app_nonce - the network server nonce.
-    /// * nwk_addr - the address of the network.
-    /// * dev_nonce - the nonce from the device.
-    /// * key - the app key.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// let dev_nonce = vec![0xcc, 0xdd];
-    /// let data = vec![
-    ///     0x20, 0x49, 0x3e, 0xeb, 0x51, 0xfb, 0xa2, 0x11, 0x6f, 0x81, 0x0e, 0xdb, 0x37, 0x42, 0x97,
-    ///     0x51, 0x42,
-    /// ];
-    /// let app_key = lorawan::keys::AppKey::from([
-    ///     0x00, 0x11, 0x22, 0x33, 0x44, 0x55, 0x66, 0x77, 0x88, 0x99, 0xaa, 0xbb, 0xcc, 0xdd, 0xee,
-    ///     0xff,
-    /// ]);
-    /// let join_accept = lorawan::parser::EncryptedJoinAcceptPayload::new(data).unwrap().decrypt(&app_key, &lorawan::default_crypto::DefaultFactory);
-    ///
-    /// let app_skey = join_accept
-    ///     .derive_appskey(&lorawan::parser::DevNonce::new(&dev_nonce[..]).unwrap(), &app_key, &lorawan::default_crypto::DefaultFactory);
-    /// ```
-    pub fn derive_appskey<TT: AsRef<[u8]>, C: CryptoFactory>(
-        &self,
-        dev_nonce: &DevNonce<TT>,
+        dev_nonce: DevNonce,
         key: &AppKey,
         crypto: &C,
     ) -> AppSKey {
-        AppSKey(self.derive_session_key(0x2, dev_nonce, &key.0, crypto))
+        AppSKey(self.derive_session_key(0x02, dev_nonce, key, crypto))
     }
 
-    fn derive_session_key<TT: AsRef<[u8]>, C: CryptoFactory>(
+    fn derive_session_key<C: CryptoFactory>(
         &self,
         first_byte: u8,
-        dev_nonce: &DevNonce<TT>,
-        key: &AES128,
+        dev_nonce: DevNonce,
+        key: &AppKey,
         crypto: &C,
     ) -> AES128 {
-        let cipher = crypto.new_enc(key);
-
-        // note: AppNonce is 24 bits, NetId is 24 bits, DevNonce is 16 bits
-        let app_nonce = self.app_nonce();
-        let nwk_addr = self.net_id();
-        let (app_nonce_arr, nwk_addr_arr, dev_nonce_arr) =
-            (app_nonce.as_ref(), nwk_addr.as_ref(), dev_nonce.as_ref());
-
+        let cipher = crypto.new_enc(key.inner());
         let mut block = [0u8; 16];
         block[0] = first_byte;
-        block[1] = app_nonce_arr[0];
-        block[2] = app_nonce_arr[1];
-        block[3] = app_nonce_arr[2];
-        block[4] = nwk_addr_arr[0];
-        block[5] = nwk_addr_arr[1];
-        block[6] = nwk_addr_arr[2];
-        block[7] = dev_nonce_arr[0];
-        block[8] = dev_nonce_arr[1];
-
+        block[1..4].copy_from_slice(self.join_nonce().as_wire_bytes());
+        block[4..7].copy_from_slice(self.net_id().as_wire_bytes());
+        block[7..9].copy_from_slice(dev_nonce.as_wire_bytes());
         cipher.encrypt_block(&mut block);
         AES128(block)
     }
 }
 
-#[derive(Debug, PartialEq, Eq)]
-pub enum CfList<'a> {
-    DynamicChannel([Frequency<'a>; 5]),
-    FixedChannel(ChannelMask<9>),
-}
+// --- Fixed-size wire fields ----------------------------------------------
+//
+// Every multi-byte LoRaWAN field is transmitted least-significant-byte
+// first. Each type here stores the bytes in wire order and offers explicit
+// conversions to and from the logical integer value, so byte order is
+// decided at exactly one place instead of at every call site.
 
-impl<T: AsRef<[u8]>> DecryptedJoinAcceptPayload<T> {
-    /// Gives the app nonce of the JoinAccept.
-    pub fn app_nonce(&self) -> AppNonce<&[u8]> {
-        const OFFSET: usize = MHDR_LEN;
-        const END: usize = OFFSET + JOIN_NONCE_LEN;
-        AppNonce::new_from_raw(&self.0.as_ref()[OFFSET..END])
-    }
+/// Error from parsing a fixed-size hex string.
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub struct ParseHexError;
 
-    /// Gives the net ID of the JoinAccept.
-    pub fn net_id(&self) -> NwkAddr<&[u8]> {
-        const OFFSET: usize = MHDR_LEN + JOIN_NONCE_LEN;
-        const END: usize = OFFSET + NET_ID_LEN;
-        NwkAddr::new_from_raw(&self.0.as_ref()[OFFSET..END])
-    }
-
-    /// Gives the dev address of the JoinAccept.
-    pub fn dev_addr(&self) -> DevAddr<&[u8]> {
-        const OFFSET: usize = MHDR_LEN + JOIN_NONCE_LEN + NET_ID_LEN;
-        const END: usize = OFFSET + DEV_ADDR_LEN;
-        DevAddr::new_from_raw(&self.0.as_ref()[OFFSET..END])
-    }
-
-    /// Gives the downlink configuration of the JoinAccept.
-    pub fn dl_settings(&self) -> DLSettings {
-        const OFFSET: usize = MHDR_LEN + JOIN_NONCE_LEN + NET_ID_LEN + DEV_ADDR_LEN;
-        DLSettings::new(self.0.as_ref()[OFFSET])
-    }
-
-    /// Gives the RX delay of the JoinAccept.
-    pub fn rx_delay(&self) -> u8 {
-        const OFFSET: usize =
-            MHDR_LEN + JOIN_NONCE_LEN + NET_ID_LEN + DEV_ADDR_LEN + DL_SETTINGS_LEN;
-        self.0.as_ref()[OFFSET] & 0x0f
-    }
-
-    /// Gives the channel frequency list of the JoinAccept.
-    pub fn c_f_list(&self) -> Option<CfList<'_>> {
-        if self.0.as_ref().len() == JOIN_ACCEPT_LEN {
-            return None;
-        }
-
-        let d = self.0.as_ref();
-        let c_f_list_type = d[28];
-        if c_f_list_type == 0 {
-            let res = [
-                Frequency::new_from_raw(&d[13..16]),
-                Frequency::new_from_raw(&d[16..19]),
-                Frequency::new_from_raw(&d[19..22]),
-                Frequency::new_from_raw(&d[22..25]),
-                Frequency::new_from_raw(&d[25..28]),
-            ];
-            Some(CfList::DynamicChannel(res))
-        } else if c_f_list_type == 1 {
-            Some(CfList::FixedChannel(ChannelMask::new_from_raw(&d[13..22])))
-        } else {
-            None
-        }
+impl core::fmt::Display for ParseHexError {
+    #[inline]
+    fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
+        f.write_str("expected a fixed-size hex string")
     }
 }
 
-impl<T: AsRef<[u8]> + AsMut<[u8]>> DecryptedJoinAcceptPayload<T> {
-    /// Creates a DecryptedJoinAcceptPayload from the bytes of a JoinAccept.
+impl core::error::Error for ParseHexError {}
+
+/// Copies a slice into a fixed array. Callers guarantee `slice.len() == N`
+/// through parse-time validation (see `Layout` in the data module and the
+/// length checks in the join module).
+#[inline]
+pub(crate) fn arr<const N: usize>(slice: &[u8]) -> [u8; N] {
+    let mut out = [0u8; N];
+    out.copy_from_slice(slice);
+    out
+}
+
+macro_rules! wire_value_newtype {
+    (
+        $(#[$outer:meta])*
+        $name:ident([u8; $n:literal], $int:ty)
+    ) => {
+        $(#[$outer])*
+        #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Default)]
+        #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
+        #[cfg_attr(feature = "defmt-03", derive(defmt::Format))]
+        pub struct $name([u8; $n]);
+
+        impl $name {
+            pub const BYTE_LEN: usize = $n;
+
+            /// Constructs from bytes in wire (transmission) order, LSB first.
+            #[inline]
+            pub const fn from_wire_bytes(bytes: [u8; $n]) -> Self {
+                Self(bytes)
+            }
+
+            /// Constructs from the logical value.
+            #[inline]
+            pub fn from_value(value: $int) -> Self {
+                let le = value.to_le_bytes();
+                let mut bytes = [0u8; $n];
+                bytes.copy_from_slice(&le[..$n]);
+                Self(bytes)
+            }
+
+            /// The bytes in wire (transmission) order, LSB first.
+            #[inline]
+            pub const fn as_wire_bytes(&self) -> &[u8; $n] {
+                &self.0
+            }
+
+            /// The logical value.
+            #[inline]
+            pub fn value(&self) -> $int {
+                let mut le = [0u8; core::mem::size_of::<$int>()];
+                le[..$n].copy_from_slice(&self.0);
+                <$int>::from_le_bytes(le)
+            }
+        }
+
+        impl From<$int> for $name {
+            #[inline]
+            fn from(value: $int) -> Self {
+                Self::from_value(value)
+            }
+        }
+
+        impl From<$name> for $int {
+            #[inline]
+            fn from(v: $name) -> Self {
+                v.value()
+            }
+        }
+
+        /// Hex-formats the logical value MSB first, the conventional display
+        /// order.
+        impl core::fmt::Display for $name {
+            #[inline]
+            fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
+                write!(f, "{:0width$x}", self.value(), width = 2 * $n)
+            }
+        }
+
+        /// Parses the conventional MSB-first hex form.
+        impl core::str::FromStr for $name {
+            type Err = ParseHexError;
+
+            #[inline]
+            fn from_str(s: &str) -> Result<Self, Self::Err> {
+                if s.len() != 2 * $n {
+                    return Err(ParseHexError);
+                }
+                let value = <$int>::from_str_radix(s, 16).map_err(|_| ParseHexError)?;
+                Ok(Self::from_value(value))
+            }
+        }
+    };
+}
+
+wire_value_newtype! {
+    /// A 32-bit LoRaWAN device address.
     ///
-    /// The JoinAccept payload is automatically decrypted and the mic is verified using the suplied
-    /// crypto factory implementation.
+    /// `DevAddr::from_value(0x01020304).as_wire_bytes()` is
+    /// `[0x04, 0x03, 0x02, 0x01]`.
+    DevAddr([u8; 4], u32)
+}
+
+impl DevAddr {
+    /// The 7-bit network identifier (NwkID) in the most significant bits.
+    #[inline]
+    pub const fn nwk_id(&self) -> u8 {
+        self.0[3] >> 1
+    }
+}
+
+wire_value_newtype! {
+    /// The 64-bit extended unique identifier of an end-device.
     ///
-    /// # Argument
+    /// Matches the conventions of [`crate::keys::DevEui`]: wire bytes are
+    /// LSB first, `Display`/`FromStr` use the MSB-first form printed on
+    /// device labels.
+    DevEui([u8; 8], u64)
+}
+
+wire_value_newtype! {
+    /// The 64-bit identifier of the join server (AppEUI before LoRaWAN
+    /// 1.0.4).
+    JoinEui([u8; 8], u64)
+}
+
+impl From<crate::keys::DevEui> for DevEui {
+    #[inline]
+    fn from(eui: crate::keys::DevEui) -> Self {
+        Self::from_wire_bytes(arr(eui.as_ref()))
+    }
+}
+
+impl From<DevEui> for crate::keys::DevEui {
+    #[inline]
+    fn from(eui: DevEui) -> Self {
+        Self::from(*eui.as_wire_bytes())
+    }
+}
+
+impl From<crate::keys::AppEui> for JoinEui {
+    #[inline]
+    fn from(eui: crate::keys::AppEui) -> Self {
+        Self::from_wire_bytes(arr(eui.as_ref()))
+    }
+}
+
+impl From<JoinEui> for crate::keys::AppEui {
+    #[inline]
+    fn from(eui: JoinEui) -> Self {
+        Self::from(*eui.as_wire_bytes())
+    }
+}
+
+wire_value_newtype! {
+    /// The 32-bit multicast group address of a Remote Multicast Setup
+    /// (TS005) session. Same wire shape as [`DevAddr`], kept distinct so the
+    /// two cannot be interchanged.
+    McAddr([u8; 4], u32)
+}
+
+wire_value_newtype! {
+    /// The 16-bit device nonce from a JoinRequest.
+    DevNonce([u8; 2], u16)
+}
+
+wire_value_newtype! {
+    /// The 24-bit server nonce from a JoinAccept.
     ///
-    /// * bytes - the data from which the PhyPayload is to be built.
-    /// * key - the key that is to be used to decrypt the payload.
-    /// * factory - the factory that shall be used to create object for crypto functions.
-    pub fn new_with_factory<C: CryptoFactory>(
-        data: T,
-        key: &AppKey,
-        crypto: &C,
-    ) -> Result<Self, Error> {
-        let t = EncryptedJoinAcceptPayload::new(data)?;
-        let res = t.decrypt(key, crypto);
-        if res.validate_mic(key, crypto) {
-            Ok(res)
-        } else {
-            Err(Error::InvalidMic)
-        }
+    /// Named AppNonce before LoRaWAN 1.0.4.
+    JoinNonce([u8; 3], u32)
+}
+
+wire_value_newtype! {
+    /// The 24-bit network identifier from a JoinAccept.
+    NetId([u8; 3], u32)
+}
+
+/// A channel frequency, stored as the raw 24-bit wire value (units of 100 Hz).
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Default)]
+#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
+#[cfg_attr(feature = "defmt-03", derive(defmt::Format))]
+pub struct Frequency([u8; 3]);
+
+impl Frequency {
+    /// Constructs from bytes in wire (transmission) order, LSB first.
+    #[inline]
+    pub const fn from_wire_bytes(bytes: [u8; 3]) -> Self {
+        Self(bytes)
+    }
+
+    /// Constructs from a frequency in Hz. Sub-100 Hz precision is lost.
+    #[inline]
+    pub fn from_hz(hz: u32) -> Self {
+        let raw = hz / 100;
+        let le = raw.to_le_bytes();
+        Self([le[0], le[1], le[2]])
+    }
+
+    /// The frequency in Hz.
+    #[inline]
+    pub fn hz(&self) -> u32 {
+        u32::from_le_bytes([self.0[0], self.0[1], self.0[2], 0]) * 100
+    }
+
+    /// The bytes in wire (transmission) order, LSB first.
+    #[inline]
+    pub const fn as_wire_bytes(&self) -> &[u8; 3] {
+        &self.0
     }
 }
 
-/// Helper trait for EncryptedDataPayload and DecryptedDataPayload.
-///
-/// NOTE: Does not check the payload size as that should be done prior to building the object of
-/// the implementing type.
-pub trait DataHeader {
-    /// Equivalent to AsRef<[u8]>.
-    fn as_data_bytes(&self) -> &[u8];
-
-    /// Gives the FHDR of the DataPayload.
-    fn fhdr(&self) -> FHDR<'_> {
-        FHDR::new_from_raw(&self.as_data_bytes()[1..(1 + self.fhdr_length())], self.is_uplink())
-    }
-
-    /// Gives whether the frame is confirmed
-    fn is_confirmed(&self) -> bool {
-        let mtype = MHDR(self.as_data_bytes()[0]).mtype();
-        mtype == MType::ConfirmedDataUp || mtype == MType::ConfirmedDataDown
-    }
-
-    /// Gives whether the payload is uplink or not.
-    fn is_uplink(&self) -> bool {
-        let mtype = MHDR(self.as_data_bytes()[0]).mtype();
-        mtype == MType::UnconfirmedDataUp || mtype == MType::ConfirmedDataUp
-    }
-
-    /// Gives the FPort of the DataPayload if there is one.
-    fn f_port(&self) -> Option<u8> {
-        let fhdr_length = self.fhdr_length();
-        let data = self.as_data_bytes();
-        if fhdr_length + 1 >= data.len() - 5 {
-            return None;
-        }
-        Some(data[1 + fhdr_length])
-    }
-
-    /// Gives the length of the FHDR field.
-    fn fhdr_length(&self) -> usize {
-        fhdr_length(self.as_data_bytes()[5])
-    }
+/// The FCtrl octet, interpreted for the frame direction it came from.
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+#[cfg_attr(feature = "defmt-03", derive(defmt::Format))]
+pub struct FCtrl {
+    byte: u8,
+    uplink: bool,
 }
-
-fn fhdr_length(b: u8) -> usize {
-    7 + (b & 0x0f) as usize
-}
-
-impl<T: DataHeader> AsPhyPayloadBytes for T {
-    fn as_bytes(&self) -> &[u8] {
-        self.as_data_bytes()
-    }
-}
-
-/// EncryptedDataPayload represents an encrypted data payload.
-///
-/// It can be built either directly through the [new](#method.new) or using the
-/// [parse](fn.parse.html) function.
-#[derive(Debug, PartialEq, Eq)]
-pub struct EncryptedDataPayload<T>(T);
-
-impl<T: AsRef<[u8]>> DataHeader for EncryptedDataPayload<T> {
-    fn as_data_bytes(&self) -> &[u8] {
-        self.0.as_ref()
-    }
-}
-
-impl<T: AsRef<[u8]>> EncryptedDataPayload<T> {
-    /// Creates a new EncryptedDataPayload if the provided data is acceptable.
-    ///
-    /// # Argument
-    ///
-    /// * data - the bytes for the payload.
-    pub fn new(data: T) -> Result<Self, Error> {
-        if Self::can_build_from(data.as_ref()) {
-            Ok(Self(data))
-        } else {
-            Err(Error::InvalidData)
-        }
-    }
-
-    fn can_build_from(bytes: &[u8]) -> bool {
-        // The smallest packets contain MHDR (1 bytes) + FHDR + MIC (4 bytes).
-        if bytes.len() < 12 || 5 + fhdr_length(bytes[5]) > bytes.len() {
-            return false;
-        }
-
-        matches!(
-            MHDR(bytes[0]).mtype(),
-            MType::ConfirmedDataUp
-                | MType::ConfirmedDataDown
-                | MType::UnconfirmedDataUp
-                | MType::UnconfirmedDataDown
-        )
-    }
-
-    /// Verifies that the DataPayload has correct MIC.
-    pub fn validate_mic<C: CryptoFactory>(&self, key: &AES128, fcnt: u32, crypto: &C) -> bool {
-        self.mic() == self.calculate_mic(key, fcnt, crypto)
-    }
-
-    fn calculate_mic<C: CryptoFactory>(&self, key: &AES128, fcnt: u32, crypto: &C) -> MIC {
-        let d = self.0.as_ref();
-        securityhelpers::calculate_data_mic(&d[..d.len() - MIC_LEN], crypto.new_mac(key), fcnt)
-    }
-}
-
-impl<T: AsRef<[u8]> + AsMut<[u8]>> EncryptedDataPayload<T> {
-    /// Decrypts the EncryptedDataPayload payload.
-    ///
-    /// This method consumes the EncryptedDataPayload as it reuses the underlying memory. Please
-    /// note that it does not verify the mic.
-    ///
-    /// If used on the application server side for application payload decryption, the nwk_skey can
-    /// be None. If used on the network server side and the app_skey is not available, app_skey can
-    /// be None when fport is 0. Failure to meet those constraints will result in an Err being
-    /// returned.
-    ///
-    /// # Argument
-    ///
-    /// * nwk_skey - the Network Session key used to decrypt the mac commands in case the payload is
-    ///   transporting those.
-    /// * app_skey - the Application Session key used to decrypt the application payload in case the
-    ///   payload is transporting that.
-    /// * fcnt - the counter used to encrypt the payload.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// let mut data = vec![
-    ///     0x40, 0x04, 0x03, 0x02, 0x01, 0x80, 0x01, 0x00, 0x01, 0xa6, 0x94, 0x64, 0x26, 0x15, 0xd6,
-    ///     0xc3, 0xb5, 0x82,
-    /// ];
-    /// let key = lorawan::keys::AES128([1; 16]);
-    /// let enc_phy = lorawan::parser::EncryptedDataPayload::new(data).unwrap();
-    /// let dec_phy = enc_phy.decrypt(None, Some(&key), 1, &lorawan::default_crypto::DefaultFactory);
-    /// ```
-    pub fn decrypt<'a, C: CryptoFactory>(
-        mut self,
-        nwk_skey: Option<&'a AES128>,
-        app_skey: Option<&'a AES128>,
-        fcnt: u32,
-        crypto: &C,
-    ) -> Result<DecryptedDataPayload<T>, Error> {
-        let fhdr_length = self.fhdr_length();
-        let fhdr = self.fhdr();
-        let full_fcnt = compute_fcnt(fcnt, fhdr.fcnt());
-        let key = if self.f_port().is_some() && self.f_port().unwrap() != 0 {
-            app_skey
-        } else {
-            nwk_skey
-        };
-        if key.is_none() {
-            return Err(Error::InvalidKey);
-        }
-        let data = self.0.as_mut();
-        let len = data.len();
-        let start = MHDR_LEN + FPORT_LEN + fhdr_length;
-        let end = len - MIC_LEN;
-        if start < end {
-            securityhelpers::encrypt_frm_data_payload(
-                data,
-                start,
-                end,
-                full_fcnt,
-                &crypto.new_enc(key.unwrap()),
-            );
-        }
-
-        Ok(DecryptedDataPayload(self.0))
-    }
-
-    /// Verifies the mic and decrypts the EncryptedDataPayload payload if mic matches.
-    ///
-    /// This is helper method that combines validate_mic and decrypt. In case the mic is fine, it
-    /// consumes the EncryptedDataPayload and reuses the underlying memory to produce
-    /// DecryptedDataPayload. If the mic does not match, it returns the original
-    /// EncryptedDataPayload so that it can be tried against the keys of another device that shares
-    /// the same dev_addr. For an example please see [decrypt](#method.decrypt).
-    pub fn decrypt_if_mic_ok<'a, C: CryptoFactory>(
-        self,
-        nwk_skey: &'a AES128,
-        app_skey: &'a AES128,
-        fcnt: u32,
-        crypto: &C,
-    ) -> Result<DecryptedDataPayload<T>, Self> {
-        if !self.validate_mic(nwk_skey, fcnt, crypto) {
-            Err(self)
-        } else {
-            Ok(self.decrypt(Some(nwk_skey), Some(app_skey), fcnt, crypto).unwrap())
-        }
-    }
-}
-
-fn compute_fcnt(old_fcnt: u32, fcnt: u16) -> u32 {
-    ((old_fcnt >> 16) << 16) ^ u32::from(fcnt)
-}
-
-/// DecryptedDataPayload represents a decrypted DataPayload.
-///
-/// It can be built either directly through the [new](#method.new) or using the
-/// [EncryptedDataPayload.decrypt](struct.EncryptedDataPayload.html#method.decrypt) function.
-#[derive(Debug, PartialEq, Eq)]
-pub struct DecryptedDataPayload<T>(T);
-
-impl<T> DecryptedDataPayload<T> {
-    pub fn to_inner(self) -> T {
-        self.0
-    }
-}
-
-#[cfg(feature = "defmt-03")]
-impl<T: AsRef<[u8]>> defmt::Format for DecryptedDataPayload<T> {
-    fn format(&self, fmt: defmt::Formatter<'_>) {
-        defmt::write!(fmt, "DecryptedDataPayload {{ {:?} }}", self.0.as_ref())
-    }
-}
-
-impl<T: AsRef<[u8]>> DataHeader for DecryptedDataPayload<T> {
-    fn as_data_bytes(&self) -> &[u8] {
-        self.0.as_ref()
-    }
-}
-
-impl<T: AsRef<[u8]>> DecryptedDataPayload<T> {
-    /// Returns FRMPayload that can represent either application payload or mac commands if fport
-    /// is 0.
-    pub fn frm_payload(&self) -> FRMPayload<'_> {
-        let data = self.as_data_bytes();
-        let len = data.len();
-        let fhdr_length = self.fhdr_length();
-        //we have more bytes than fhdr + fport
-        if len < fhdr_length + 6 {
-            FRMPayload::None
-        } else if self.f_port() != Some(0) {
-            // the size guarantees the existance of f_port
-            FRMPayload::Data(&data[(1 + fhdr_length + 1)..(len - 4)])
-        } else {
-            FRMPayload::MACCommands(FRMMacCommands::new(
-                &data[(1 + fhdr_length + 1)..(len - 4)],
-                self.is_uplink(),
-            ))
-        }
-    }
-}
-
-/// Parses a payload as LoRaWAN physical payload.
-///
-/// # Argument
-///
-/// * bytes - the data from which the PhyPayload is to be built.
-///
-/// # Examples
-///
-/// ```
-/// let mut data = vec![
-///     0x40, 0x04, 0x03, 0x02, 0x01, 0x80, 0x01, 0x00, 0x01, 0xa6, 0x94, 0x64, 0x26, 0x15, 0xd6,
-///     0xc3, 0xb5, 0x82,
-/// ];
-/// if let Ok(lorawan::parser::PhyPayload::Data(phy)) = lorawan::parser::parse(data) {
-///     println!("{:?}", phy);
-/// } else {
-///     panic!("failed to parse data payload");
-/// }
-/// ```
-pub fn parse<T>(data: T) -> Result<PhyPayload<T>, Error>
-where
-    T: AsRef<[u8]> + AsMut<[u8]>,
-{
-    let bytes = data.as_ref();
-    // Enough data for payload (bare minimum without fport and FRMPayload)?
-    if bytes.len() < PHY_PAYLOAD_MIN_LEN {
-        return Err(Error::InvalidPayload);
-    }
-    let mhdr = MHDR(bytes[0]);
-    if mhdr.major() != Major::LoRaWANR1 {
-        return Err(Error::UnsupportedMajorVersion);
-    }
-    match mhdr.mtype() {
-        MType::JoinRequest => Ok(PhyPayload::JoinRequest(JoinRequestPayload::new(data)?)),
-        MType::JoinAccept => Ok(PhyPayload::JoinAccept(JoinAcceptPayload::Encrypted(
-            EncryptedJoinAcceptPayload::new(data)?,
-        ))),
-        MType::UnconfirmedDataUp
-        | MType::ConfirmedDataUp
-        | MType::UnconfirmedDataDown
-        | MType::ConfirmedDataDown => {
-            Ok(PhyPayload::Data(DataPayload::Encrypted(EncryptedDataPayload::new(data)?)))
-        }
-        _ => Err(Error::InvalidMessageType),
-    }
-}
-
-/// MHDR represents LoRaWAN MHDR.
-#[derive(Debug, PartialEq, Eq)]
-pub struct MHDR(u8);
-
-impl MHDR {
-    pub fn new(byte: u8) -> MHDR {
-        MHDR(byte)
-    }
-
-    /// Type of message PhyPayload is carrying.
-    pub fn mtype(&self) -> MType {
-        match self.0 >> 5 {
-            0 => MType::JoinRequest,
-            1 => MType::JoinAccept,
-            2 => MType::UnconfirmedDataUp,
-            3 => MType::UnconfirmedDataDown,
-            4 => MType::ConfirmedDataUp,
-            5 => MType::ConfirmedDataDown,
-            6 => MType::RFU,
-            _ => MType::Proprietary,
-        }
-    }
-
-    /// Version of LoRaWAN payload format.
-    pub fn major(&self) -> Major {
-        if self.0.trailing_zeros() >= 2 {
-            Major::LoRaWANR1
-        } else {
-            Major::RFU
-        }
-    }
-}
-
-impl From<u8> for MHDR {
-    fn from(v: u8) -> Self {
-        MHDR(v)
-    }
-}
-
-/// MType gives the possible message types of the PhyPayload.
-#[derive(Debug, PartialEq, Eq)]
-pub enum MType {
-    JoinRequest,
-    JoinAccept,
-    UnconfirmedDataUp,
-    UnconfirmedDataDown,
-    ConfirmedDataUp,
-    ConfirmedDataDown,
-    RFU,
-    Proprietary,
-}
-
-/// Major gives the supported LoRaWAN payload formats.
-#[derive(Debug, PartialEq, Eq)]
-pub enum Major {
-    LoRaWANR1,
-    RFU,
-}
-
-fixed_len_struct! {
-    /// EUI64 represents a 64-bit EUI.
-    struct EUI64[8];
-}
-
-fixed_len_struct! {
-    /// DevNonce represents a 16-bit device nonce.
-    struct DevNonce[2];
-}
-
-impl From<DevNonce<[u8; 2]>> for u16 {
-    fn from(v: DevNonce<[u8; 2]>) -> Self {
-        u16::from_be_bytes(v.0)
-    }
-}
-
-impl From<u16> for DevNonce<[u8; 2]> {
-    fn from(v: u16) -> Self {
-        Self::from(v.to_be_bytes())
-    }
-}
-
-fixed_len_struct! {
-    /// AppNonce represents a 24-bit network server nonce.
-    struct AppNonce[3];
-}
-
-fixed_len_struct! {
-    /// DevAddr represents a 32-bit device address.
-    struct DevAddr[4];
-}
-
-#[allow(clippy::should_implement_trait)]
-impl<T: AsRef<[u8]>> DevAddr<T> {
-    pub fn nwk_id(&self) -> u8 {
-        self.0.as_ref()[0] >> 1
-    }
-    pub fn as_ref(&self) -> &[u8] {
-        self.0.as_ref()
-    }
-}
-
-impl From<DevAddr<[u8; 4]>> for u32 {
-    fn from(v: DevAddr<[u8; 4]>) -> Self {
-        u32::from_be_bytes(v.0)
-    }
-}
-
-impl From<u32> for DevAddr<[u8; 4]> {
-    fn from(v: u32) -> Self {
-        Self::from(v.to_be_bytes())
-    }
-}
-
-fixed_len_struct! {
-    /// `McAddr` represents a 32-bit multicast address.
-    struct McAddr[4];
-}
-
-impl From<McAddr<[u8; 4]>> for u32 {
-    fn from(v: McAddr<[u8; 4]>) -> Self {
-        u32::from_be_bytes(v.0)
-    }
-}
-
-impl From<u32> for McAddr<[u8; 4]> {
-    fn from(v: u32) -> Self {
-        Self::from(v.to_be_bytes())
-    }
-}
-
-impl From<DevAddr<[u8; 4]>> for McAddr<[u8; 4]> {
-    fn from(v: DevAddr<[u8; 4]>) -> Self {
-        McAddr(v.0)
-    }
-}
-
-fixed_len_struct! {
-    /// NwkAddr represents a 24-bit network address.
-    struct NwkAddr[3];
-}
-
-/// FHDR represents FHDR from DataPayload.
-#[derive(Debug, PartialEq, Eq)]
-pub struct FHDR<'a>(&'a [u8], bool);
-
-impl<'a> FHDR<'a> {
-    pub fn new_from_raw(bytes: &'a [u8], uplink: bool) -> FHDR<'a> {
-        FHDR(bytes, uplink)
-    }
-
-    pub fn new(bytes: &'a [u8], uplink: bool) -> Option<FHDR<'a>> {
-        let data_len = bytes.len();
-        if data_len < 7 {
-            return None;
-        }
-        if data_len < fhdr_length(bytes[4]) {
-            return None;
-        }
-        Some(FHDR(bytes, uplink))
-    }
-
-    /// Gives the device address associated with the given payload.
-    pub fn dev_addr(&self) -> DevAddr<&'a [u8]> {
-        DevAddr::new_from_raw(&self.0[0..4])
-    }
-
-    /// Gives the multicast address associated with the given payload.
-    pub fn mc_addr(&self) -> McAddr<&'a [u8]> {
-        McAddr::new_from_raw(&self.0[0..4])
-    }
-
-    /// Gives the FCtrl associated with the given payload.
-    pub fn fctrl(&self) -> FCtrl {
-        FCtrl(self.0[4], self.1)
-    }
-
-    /// Gives the truncated FCnt associated with the given payload.
-    pub fn fcnt(&self) -> u16 {
-        (u16::from(self.0[6]) << 8) | u16::from(self.0[5])
-    }
-
-    /// Gives the size of FOpts.
-    pub fn fopts_len(&self) -> u8 {
-        FCtrl(self.0[4], self.1).f_opts_len()
-    }
-
-    pub fn data(&self) -> &[u8] {
-        &self.0[7_usize..(7 + self.fopts_len()) as usize]
-    }
-}
-
-/// FCtrl represents the FCtrl from FHDR.
-#[derive(Debug, PartialEq, Eq)]
-pub struct FCtrl(pub u8, pub bool);
 
 impl FCtrl {
-    pub fn set_ack(&mut self) {
-        self.0 |= 1 << 5;
+    #[inline]
+    pub const fn new(byte: u8, uplink: bool) -> Self {
+        Self { byte, uplink }
     }
 
-    pub fn new(bytes: u8, uplink: bool) -> FCtrl {
-        FCtrl(bytes, uplink)
+    #[inline]
+    pub const fn adr(&self) -> bool {
+        self.byte & 0x80 != 0
     }
 
-    /// Set ADR ACK requested.
-    pub fn set_adr_ack_req(&mut self) {
-        self.0 |= 1 << 6;
+    /// ADR ACK requested; uplink only.
+    #[inline]
+    pub const fn adr_ack_req(&self) -> bool {
+        self.uplink && self.byte & 0x40 != 0
     }
 
-    /// Set ADR enabled.
-    pub fn set_adr(&mut self) {
-        self.0 |= 1 << 7;
+    #[inline]
+    pub const fn ack(&self) -> bool {
+        self.byte & 0x20 != 0
     }
 
-    /// Gives whether ADR is enabled or not.
-    pub fn adr(&self) -> bool {
-        self.0 >> 7 == 1
+    /// More downlink pending; downlink only.
+    #[inline]
+    pub const fn f_pending(&self) -> bool {
+        !self.uplink && self.byte & 0x10 != 0
     }
 
-    /// Gives whether ADR ACK is requested.
-    pub fn adr_ack_req(&self) -> bool {
-        self.1 && self.0 & (1 << 6) != 0
+    #[inline]
+    pub const fn f_opts_len(&self) -> usize {
+        (self.byte & 0x0f) as usize
     }
 
-    /// Gives whether ack bit is set.
-    pub fn ack(&self) -> bool {
-        self.0 & (1 << 5) != 0
+    #[inline]
+    pub const fn raw_value(&self) -> u8 {
+        self.byte
     }
-
-    /// Gives whether there are more payloads pending.
-    pub fn f_pending(&self) -> bool {
-        !self.1 && self.0 & (1 << 4) != 0
-    }
-
-    /// Gives the size of FOpts.
-    pub fn f_opts_len(&self) -> u8 {
-        self.0 & 0x0f
-    }
-
-    /// Gives the binary representation of the FCtrl.
-    pub fn raw_value(&self) -> u8 {
-        self.0
-    }
-}
-
-/// FRMPayload represents the FRMPayload that can either be the application
-/// data or mac commands.
-#[derive(Debug, PartialEq, Eq)]
-pub enum FRMPayload<'a> {
-    Data(&'a [u8]),
-    MACCommands(FRMMacCommands<'a>),
-    None,
-}
-
-/// FRMMacCommands represents the mac commands.
-#[derive(Debug, PartialEq, Eq)]
-pub struct FRMMacCommands<'a>(bool, &'a [u8]);
-
-impl<'a> FRMMacCommands<'a> {
-    pub fn new(bytes: &'a [u8], uplink: bool) -> Self {
-        FRMMacCommands(uplink, bytes)
-    }
-
-    pub fn data(&self) -> &[u8] {
-        self.1
-    }
-
-    // /// Gives the list of mac commands represented in the FRMPayload.
-    // pub fn mac_commands(&self) -> MacCommandIterator {
-    //     parse_mac_commands(self.1, self.0)
-    // }
 }
