--- lorawan-encoding/src/v2/mac.rs
@@ -1,103 +1,0 @@
-//! MAC command parsing with explicit error signaling.
-//!
-//! The old `MacCommandIterator` yields plain commands and simply stops at the
-//! first unknown CID or truncated payload, so a malformed tail is
-//! indistinguishable from a clean end of stream (its docs promise an error it
-//! cannot return), and for variable-length commands an empty payload can
-//! panic (e.g. multicast `McGroupStatusAns` parsed from a lone CID byte).
-//! This iterator yields `Result` and fuses after the first error: garbage is
-//! reported exactly once, then iteration ends.
-//!
-//! The [`MacCommandSet`] impls are generated by the `CommandHandler` derive
-//! from the same `#[cmd(cid = ..., len = ...)]` attributes that drive the
-//! legacy iterator, so the two framings can never drift apart. The command
-//! payload types (accessors, wire layout) are shared with
-//! [`crate::maccommands`], [`crate::certification`] and [`crate::multicast`];
-//! only the framing layer differs.
-
-use core::marker::PhantomData;
-
-use crate::maccommands::{DownlinkMacCommand, UplinkMacCommand};
-
-/// Error from parsing a MAC command stream.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-#[cfg_attr(feature = "defmt-03", derive(defmt::Format))]
-pub enum Error {
-    /// A CID that is not defined for this command set.
-    UnknownCid(u8),
-    /// The stream ends in the middle of this command's payload.
-    Truncated { cid: u8 },
-}
-
-impl core::fmt::Display for Error {
-    fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
-        match self {
-            Error::UnknownCid(cid) => write!(f, "unknown MAC command CID 0x{cid:02x}"),
-            Error::Truncated { cid } => {
-                write!(f, "truncated payload for MAC command CID 0x{cid:02x}")
-            }
-        }
-    }
-}
-
-impl core::error::Error for Error {}
-
-/// One command set (a direction of a package): parses a single command off
-/// the front of a byte stream.
-///
-/// Implementations are generated by the `CommandHandler` derive macro from
-/// the enum's `#[cmd]` attributes.
-pub trait MacCommandSet<'a>: Sized {
-    /// Parses one command from the start of `data` (guaranteed non-empty by
-    /// the caller), returning the command and the number of bytes consumed.
-    fn parse_one(data: &'a [u8]) -> Result<(Self, usize), Error>;
-}
-
-/// Iterator over a MAC command stream, yielding `Result` per command.
-///
-/// Fused after the first error: a malformed tail yields exactly one `Err`.
-#[derive(Debug, Clone)]
-pub struct MacCommands<'a, T> {
-    data: &'a [u8],
-    errored: bool,
-    _commands: PhantomData<T>,
-}
-
-impl<'a, T> MacCommands<'a, T> {
-    #[inline]
-    pub fn new(data: &'a [u8]) -> Self {
-        Self { data, errored: false, _commands: PhantomData }
-    }
-}
-
-impl<'a, T: MacCommandSet<'a>> Iterator for MacCommands<'a, T> {
-    type Item = Result<T, Error>;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        if self.errored || self.data.is_empty() {
-            return None;
-        }
-        match T::parse_one(self.data) {
-            Ok((cmd, consumed)) => {
-                self.data = &self.data[consumed..];
-                Some(Ok(cmd))
-            }
-            Err(e) => {
-                self.errored = true;
-                Some(Err(e))
-            }
-        }
-    }
-}
-
-/// Parses a stream of uplink (device-transmitted) MAC commands.
-#[inline]
-pub fn parse_uplink_mac_commands(data: &[u8]) -> MacCommands<'_, UplinkMacCommand<'_>> {
-    MacCommands::new(data)
-}
-
-/// Parses a stream of downlink (server-transmitted) MAC commands.
-#[inline]
-pub fn parse_downlink_mac_commands(data: &[u8]) -> MacCommands<'_, DownlinkMacCommand<'_>> {
-    MacCommands::new(data)
-}
