--- lorawan-encoding/src/v2/mod.rs
@@ -1,188 +1,0 @@
-//! Prototype of a borrowed-view redesign for this crate (design/encoding-v2).
-//!
-//! This module demonstrates the proposed replacement for the `T: AsRef<[u8]>`
-//! generic-over-storage design:
-//!
-//! * Fixed-size wire fields ([`DevAddr`], [`DevEui`], [`JoinEui`],
-//!   [`DevNonce`], [`JoinNonce`], [`NetId`], [`Frequency`]) become plain
-//!   owned `[u8; N]` newtypes (`Copy`, `Eq`, `Hash`) with explicit
-//!   wire-order/value conversions.
-//! * Parsing produces zero-copy read-only views with all structural
-//!   validation done once at the boundary, so every accessor afterwards is
-//!   infallible by construction. One monomorphization regardless of where the
-//!   bytes live.
-//! * Decryption stays in place but takes the `&mut [u8]` explicitly, keeping
-//!   the encrypted/decrypted typestate without any storage generic.
-//! * MAC command parsing ([`mac`], [`certification`], [`multicast`]) yields
-//!   `Result` per command instead of silently stopping at malformed input,
-//!   with the framing generated by the `CommandHandler` derive from the same
-//!   `#[cmd]` attributes as the legacy iterator.
-//! * Frames are built from plain structs ([`DataFrame`], [`JoinRequest`],
-//!   [`JoinAccept`]) with `build_into(&mut buf)`; the types make most
-//!   invalid frames unrepresentable.
-//! * One [`Error`] enum implementing `Display` and `core::error::Error`.
-//!
-//! Known deliberate divergences from the `parser` module:
-//!
-//! * `f_port()` returns `Some` when an FPort byte is present with an empty
-//!   FRMPayload. The old code reports `None` in that case, conflating "FPort
-//!   absent" with "FPort present, payload empty".
-//! * Wire fields are explicit about byte order: the wire transmits them
-//!   little-endian, so `from_value`/`value` convert accordingly. The old
-//!   `From<u32> for DevAddr` and `From<u16> for DevNonce` used big-endian
-//!   array order, which round-trips through the wire with flipped endianness.
-//! * All parse entry points reject frames with a non-R1 major version; the
-//!   old `EncryptedDataPayload::new` accepted them.
-//! * A MAC command stream with a malformed tail yields `Err`; the old
-//!   iterator ends silently.
-//!
-//! # Examples
-//!
-//! Receiving on a device: inspect the frame read-only, look up the session
-//! by [`DevAddr`], verify the MIC, then decrypt in place.
-//!
-//! ```
-//! use lorawan::default_crypto::DefaultFactory;
-//! use lorawan::keys::{AppSKey, NwkSKey};
-//! use lorawan::v2::{DecryptedDataPayload, DevAddr, EncryptedDataPayload, FrmPayload};
-//!
-//! let mut buf = [
-//!     0x40, 0x04, 0x03, 0x02, 0x01, 0x80, 0x01, 0x00, 0x01, 0xa6, 0x94, 0x64, 0x26, 0x15,
-//!     0xd6, 0xc3, 0xb5, 0x82,
-//! ];
-//! let nwk_skey = NwkSKey::from([2; 16]);
-//! let app_skey = AppSKey::from([1; 16]);
-//!
-//! let phy = EncryptedDataPayload::parse(&buf).unwrap();
-//! assert_eq!(phy.fhdr().dev_addr(), DevAddr::from_value(0x01020304));
-//! assert!(phy.validate_mic(nwk_skey.inner(), 1, &DefaultFactory));
-//!
-//! let decrypted = DecryptedDataPayload::decrypt_in_place(
-//!     &mut buf,
-//!     Some(&nwk_skey),
-//!     Some(&app_skey),
-//!     1,
-//!     &DefaultFactory,
-//! )
-//! .unwrap();
-//! assert_eq!(decrypted.frm_payload(), FrmPayload::Data(b"hello"));
-//! ```
-//!
-//! Sending: describe the frame with a struct literal and build it into a
-//! caller-provided buffer.
-//!
-//! ```
-//! use core::num::NonZeroU8;
-//! use lorawan::default_crypto::DefaultFactory;
-//! use lorawan::keys::{AppSKey, NwkSKey};
-//! use lorawan::v2::{DataFrame, DataFrameType, DevAddr, Payload};
-//!
-//! let frame = DataFrame {
-//!     frame_type: DataFrameType::ConfirmedUp,
-//!     dev_addr: DevAddr::from_value(0x01020304),
-//!     adr: true,
-//!     fcnt: 42,
-//!     payload: Payload::Data { f_port: NonZeroU8::new(1).unwrap(), data: b"hello" },
-//!     ..Default::default()
-//! };
-//! let mut buf = [0u8; 64];
-//! let bytes = frame
-//!     .build_into(&mut buf, &NwkSKey::from([2; 16]), Some(&AppSKey::from([1; 16])), &DefaultFactory)
-//!     .unwrap();
-//! assert_eq!(bytes.len(), 18);
-//! ```
-
-pub mod certification;
-mod creator;
-mod data;
-mod join;
-pub mod mac;
-pub mod multicast;
-mod types;
-
-pub use creator::{DataFrame, JoinAccept, JoinRequest, Payload};
-pub use data::{DataFrameType, DecryptedDataPayload, EncryptedDataPayload, Fhdr, FrmPayload};
-pub use join::{
-    CfList, DecryptedJoinAcceptPayload, EncryptedJoinAcceptPayload, JoinRequestPayload,
-};
-pub use types::{
-    DevAddr, DevEui, DevNonce, FCtrl, Frequency, JoinEui, JoinNonce, NetId, ParseHexError,
-};
-
-/// Errors from parsing or decrypting a LoRaWAN frame.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-#[cfg_attr(feature = "defmt-03", derive(defmt::Format))]
-pub enum Error {
-    /// The buffer is shorter than the smallest possible frame.
-    TooShort,
-    /// The MHDR major version is not LoRaWAN R1.
-    UnsupportedMajorVersion,
-    /// The MHDR message type is RFU or Proprietary.
-    UnsupportedMessageType,
-    /// The MHDR message type does not match the frame type being parsed.
-    UnexpectedMessageType,
-    /// The MHDR message type is not one of the four data frame types.
-    NotADataFrame,
-    /// The buffer length is not valid for this frame type.
-    InvalidLength,
-    /// FOptsLen claims more FOpts bytes than the frame contains.
-    TruncatedFhdr,
-    /// No key was supplied for the FPort carried by this frame.
-    MissingKey,
-    /// The MIC does not match the frame contents.
-    InvalidMic,
-    /// The output buffer is too small for the frame being built.
-    BufferTooShort,
-    /// FOpts is limited to 15 bytes.
-    FOptsTooLong,
-    /// The spec forbids FPort 0 (MAC commands in FRMPayload) together with a
-    /// non-empty FOpts.
-    FOptsWithFPortZero,
-}
-
-impl core::fmt::Display for Error {
-    fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
-        let s = match self {
-            Error::BufferTooShort => "output buffer too small for the frame",
-            Error::FOptsTooLong => "FOpts is limited to 15 bytes",
-            Error::FOptsWithFPortZero => "FPort 0 with non-empty FOpts is forbidden",
-            Error::TooShort => "buffer shorter than minimal frame",
-            Error::UnsupportedMajorVersion => "unsupported LoRaWAN major version",
-            Error::UnsupportedMessageType => "RFU or proprietary MHDR message type",
-            Error::UnexpectedMessageType => "MHDR message type does not match frame type",
-            Error::NotADataFrame => "MHDR message type is not a data frame",
-            Error::InvalidLength => "buffer length invalid for this frame type",
-            Error::TruncatedFhdr => "FOptsLen exceeds frame length",
-            Error::MissingKey => "no key supplied for this frame's FPort",
-            Error::InvalidMic => "MIC mismatch",
-        };
-        f.write_str(s)
-    }
-}
-
-impl core::error::Error for Error {}
-
-/// A parsed LoRaWAN physical payload: a read-only view classified by MHDR.
-#[derive(Debug, Clone, Copy)]
-pub enum PhyPayload<'a> {
-    JoinRequest(JoinRequestPayload<'a>),
-    JoinAccept(EncryptedJoinAcceptPayload<'a>),
-    Data(EncryptedDataPayload<'a>),
-}
-
-/// Parses and classifies a LoRaWAN physical payload.
-#[inline]
-pub fn parse(bytes: &[u8]) -> Result<PhyPayload<'_>, Error> {
-    let [mhdr, ..] = bytes else {
-        return Err(Error::TooShort);
-    };
-    if mhdr & 0b11 != 0 {
-        return Err(Error::UnsupportedMajorVersion);
-    }
-    match mhdr >> 5 {
-        0 => Ok(PhyPayload::JoinRequest(JoinRequestPayload::parse(bytes)?)),
-        1 => Ok(PhyPayload::JoinAccept(EncryptedJoinAcceptPayload::parse(bytes)?)),
-        2..=5 => Ok(PhyPayload::Data(EncryptedDataPayload::parse(bytes)?)),
-        _ => Err(Error::UnsupportedMessageType),
-    }
-}
